% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFamily.R
\name{addFamily}
\alias{addFamily}
\title{A function to calculate evolutionary distinctiveness among phyloregions.}
\usage{
addFamily(comm, tpl, syn)
}
\arguments{
\item{P}{A site × site phylogenetic beta diversity distance matrix}

\item{\dots}{Further arguments passed to or from other methods.}
}
\value{
\item{addFamily}{A dataframe with evolutionary distinctiveness of each
phyloregion}
}
\description{
The \code{addFamily} function estimates evolutionary distinctiveness of
each
phyloregion by computing the mean value of phylogenetic beta diversity
between a focal phyloregion and all other phyloregions in the study area.
}
\examples{
example(phylo_beta)
addFamily(P)

}
\references{
\insertRef{daru2017understanding}{bioregion}

\insertRef{Holt2013}{bioregion}
}
\seealso{
\code{\link[picante]{evol.distinct}}
}
\author{
Barnabas H. Daru \email{darunabas@gmail.com} & Klaus Schliep
}
\keyword{cluster}
