% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{benthicdf}
\alias{benthicdf}
\title{Spatial distribution of benthic invertebrates in Brittany (data.frame)}
\format{
A \code{data.frame} with 28,563 rows and 3 columns:
\describe{
\item{Site}{Unique site identifier (corresponding to the field ID of benthicsp).}
\item{Species}{Unique species identifier.}
\item{Abundance}{Species abundance}
}
}
\usage{
benthicdf
}
\description{
A dataset containing the abundance of 3,697 species in 715 sites.
}
\keyword{datasets}
