% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{benthicmat}
\alias{benthicmat}
\title{Spatial distribution of benthic invertebrates in Brittany
(co-occurrence matrix)}
\format{
A co-occurrence \code{matrix} with sites as rows and species as
columns. Each element of the matrix
represents the abundance of the species in the site.
}
\usage{
benthicmat
}
\description{
A dataset containing the abundance of each of the 838 species in each of
the 162 sites.
}
\keyword{datasets}
