% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{benthicsp}
\alias{benthicsp}
\title{Spatial distribution of benthic invertebrates in Brittany (spatial points)}
\format{
A
\describe{
\item{ID}{Unique site identifier.}
\item{geometry}{Geometry of the site.}
}
}
\usage{
benthicsp
}
\description{
A dataset containing the geometry of the 162 sites.
}
\keyword{datasets}
