% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin}
\alias{bin}
\title{Download, unzip and check permission of bioRgeo's executables}
\usage{
bin(binpath = NULL)
}
\arguments{
\item{binpath}{a string indicating the path to the folder that will host the bin folder (bioRgeo's package by default,
if you use a different folder please be sure to update the \code{binpath} in \link{infomap}, \link{louvain}
and \link{oslom})}
}
\description{
This function download and unzip the bin folder needed to run some bioRgeo's functions. It also checks if the files
have the permissions to be executed as programs.
}
\author{
Pierre Denelle (\email{pierre.denelle@gmail.com}),
Maxime Lenormand (\email{maxime.lenormand@inrae.fr}) and
Boris Leroy (\email{leroy.boris@gmail.com})
}
