% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioregion_metrics.R
\name{bioregion_metrics}
\alias{bioregion_metrics}
\title{Calculate metrics for bioregions}
\usage{
bioregion_metrics(bioregionalization, comat, map = NULL, col_bioregion = NULL)
}
\arguments{
\item{bioregionalization}{A \code{bioregion.clusters} object.}

\item{comat}{A site-species \code{matrix} with sites as rows and species as
columns.}

\item{map}{A spatial object that can be handled by \code{sf} or \code{terra}.
The first attribute or layer should correspond to the sites' ID
(see Details). Needed only for the spatial coherence (\code{NULL} by default).}

\item{col_bioregion}{Deprecated.}
}
\value{
A \code{data.frame} with 5 columns (Bioregion ID and metrics described below) or
7 if spatial coherence is computed.
\itemize{
\item \strong{n_sites}: Number of sites per bioregion
\item \strong{richness}: Number of distinct species per bioregion.
\item \strong{rich_endemics}: Number of species found only in the bioregion.
\item \strong{prop_endemics}: Fraction of endemics species.
\item \href{https://biorgeo.github.io/bioregion/articles/a5_2_summary_metrics.html#bioregion-metrics-spatial-coherence}{\strong{sc_size}}:
Spatial coherence based on size, fraction of the number of site contained in the bioregion's largest contiguous patch.
\item \href{https://biorgeo.github.io/bioregion/articles/a5_2_summary_metrics.html#bioregion-metrics-spatial-coherence}{\strong{sc_area}}:
Spatial coherence based on area, fraction of the bioregion area contained in its largest contiguous patch.
}

Note that if \code{bioregionalization} contains multiple partitions
(i.e., if \code{dim(bioregionalization$clusters) > 2}), a list will be
returned.
}
\description{
This function calculates the number of sites, the number of species,
the number of endemic species and the proportion of endemism per bioregion. The
spatial coherence can be optionally computed if a spatial object is provided.
}
\details{
\code{map} should be the output of
\code{map_bioregions(bioregionalization,    geometry, write_clusters = TRUE)}
}
\examples{
comat <- matrix(sample(1000, 50), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

net <- similarity(comat, metric = "Simpson")
clust <- netclu_louvain(net)

bioregion_metrics(bioregionalization = clust, 
                  comat = comat) 

}
\seealso{
For more details illustrated with a practical example,
see the vignette:
\url{https://biorgeo.github.io/bioregion/articles/a5_2_summary_metrics.html#bioregion-metrics-spatial-coherence}.

Associated functions:
\link{site_species_metrics} \link{bioregionalization_metrics}
}
\author{
Maxime Lenormand (\email{maxime.lenormand@inrae.fr}) \cr
Pierre Denelle (\email{pierre.denelle@gmail.com}) \cr
Boris Leroy (\email{leroy.boris@gmail.com})
}
