% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioregionalization_metrics.R
\name{bioregionalization_metrics}
\alias{bioregionalization_metrics}
\title{Calculate metrics for a bioregionalization}
\usage{
bioregionalization_metrics(
  bioregionalization,
  eval_metrics = c("prop_between_dissim", "anosim"),
  dissimilarity,
  dissimilarity_index = names(dissimilarity)[3],
  comat = NULL,
  anosim_permutations = 1,
  eval_metric = NULL,
  net = NULL,
  site_col = NULL,
  species_col = NULL
)
}
\arguments{
\item{bioregionalization}{A \code{bioregion.clusters} object.}

\item{eval_metrics}{A \code{character} vector or a single \code{character} string
indicating the metric(s) to be calculated. Available options are
\code{"prop_between_dissim"}, \code{"anosim"}, \code{"mean_endemics"} or \code{"tot_endemics"}. Use \code{"all"}
to compute all available metrics. See Details for metric descriptions.}

\item{dissimilarity}{A site-by-site dissimilarity object from \code{\link[=dissimilarity]{dissimilarity()}}
or \code{\link[=dissimilarity_to_similarity]{dissimilarity_to_similarity()}}. Required only for \code{"prop_between_dissim"}
and \code{"anosim"}.}

\item{dissimilarity_index}{The name or number of the column to use as
dissimilarity. By default, the third column name of \code{dissimilarity} is used.}

\item{comat}{A site-species \code{matrix} with sites as rows and species as
columns. Should be provided if \code{eval_metrics} includes \code{"avg_endemism"} or
\code{"tot_endemism"}.}

\item{anosim_permutations}{The number of permutations used to compute the
p-value associated with the ANOSIM statistic. Defaults to 1, in which
case no p-value is returned.}

\item{eval_metric}{Deprecated.}

\item{net}{Deprecated.}

\item{site_col}{Deprecated.}

\item{species_col}{Deprecated.}
}
\value{
A \code{data.frame} containing the \code{eval_metrics} values for a
bioregionalization and its partition(s).
}
\description{
This function calculates metrics at the bioregionalization level. These
evaluation metrics can be used to assess and select a partition based on the
optimal number of clusters.
}
\details{
\strong{Evaluation metrics:}

\itemize{

\item{\href{https://biorgeo.github.io/bioregion/articles/a5_2_summary_metrics.html#prop_between_dissim}{\strong{prop_between_dissim}}:
The proportion of total dissimilarity occurring between
bioregions, following Holt et al. (2013), calculated as the sum of
between-bioregion dissimilarities divided by the total sum of
dissimilarities.}

\item{\href{https://biorgeo.github.io/bioregion/articles/a5_2_summary_metrics.html#anosim}{\strong{anosim}}:
The Analysis of Similarities (ANOSIM) statistic, based on
\link[vegan:anosim]{vegan::anosim()}. It measures the separation between
within- and between-bioregion dissimilarities. When \code{permutation > 1},
a p-value is calculated using permutation testing.}

\item{\href{https://biorgeo.github.io/bioregion/articles/a5_2_summary_metrics.html#mean_endemics}{\strong{mean_endemics}}:
The mean proportion of endemic species across bioregions,
following Kreft & Jetz (2010). For each bioregion, the proportion of endemic
species is calculated and then averaged across bioregions.}

\item{\href{https://biorgeo.github.io/bioregion/articles/a5_2_summary_metrics.html#tot_endemics}{\strong{tot_endemics}}:
The proportion of endemic species across all bioregion.
It is calculated as the total number of endemic species divided by the total
number of species. Endemic species are those occurring in only one bioregion.}

}
}
\examples{
comat <- matrix(sample(0:1000, size = 500, replace = TRUE, prob = 1/1:1001),
20, 25)
rownames(comat) <- paste0("Site",1:20)
colnames(comat) <- paste0("Species",1:25)

dissim <- dissimilarity(comat, metric = "all")

# User-defined number of clusters
bioreg <- hclu_hierarclust(dissim, 
                           n_clust = 10:15, 
                           index = "Simpson",
                           verbose = FALSE)

met <- bioregionalization_metrics(bioreg,
                                  eval_metrics = "all",
                                  dissimilarity = dissim,
                                  comat = comat)
met

}
\references{
Holt BG, Lessard J, Borregaard MK, Fritz SA, Araújo MB, Dimitrov D, Fabre P,
Graham CH, Graves GR, Jønsson Ka, Nogués-Bravo D, Wang Z, Whittaker RJ,
Fjeldså J & Rahbek C (2013) An update of Wallace's zoogeographic regions of
the world. \emph{Science} 339, 74-78.

Kreft H & Jetz W (2010) A framework for delineating biogeographical regions
based on species distributions. \emph{Journal of Biogeography} 37, 2029-2053.
}
\seealso{
For more details illustrated with a practical example,
see the vignette:
\url{https://biorgeo.github.io/bioregion/articles/a5_2_summary_metrics.html#bioregionalization}.

Associated functions:
\link{compare_bioregionalizations} \link{find_optimal_n}
}
\author{
Boris Leroy (\email{leroy.boris@gmail.com}) \cr
Maxime Lenormand (\email{maxime.lenormand@inrae.fr}) \cr
Pierre Denelle (\email{pierre.denelle@gmail.com})
}
