% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth.R
\name{choropleth}
\alias{choropleth}
\title{Binning shapefile polygons based upon a slot value}
\usage{
choropleth(x, k = 10, ...)
}
\arguments{
\item{\dots}{Further arguments passed to or from other methods.}

\item{P}{A site × site phylogenetic beta diversity distance matrix}
}
\value{
\item{choropleth}{A dataframe with evolutionary distinctiveness of each
phyloregion}
}
\description{
The \code{choropleth} function discretizes the values of a quantity based 
on their quantiles.
}
\examples{
library(raster)
file <- paste0(system.file(package="bioregion"), "/inst/SppRich_NG/SR_Nigeria.shp")
s <- shapefile(file)
k=10
COLOUR <- terrain.colors(k)
y = choropleth(s, k)
plot(y, col=COLOUR[y$SR])

}
\keyword{cluster}
