% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_hierarchical.R
\name{clustering_hierarchical}
\alias{clustering_hierarchical}
\title{Hierarchical clustering based distances or beta-diversity}
\usage{
clustering_hierarchical(
  distances,
  index = names(distances)[3],
  method = "average",
  randomize = TRUE,
  n_runs = 30,
  keep_trials = FALSE,
  optimal_tree_method = "best",
  n_clust = NULL,
  cut_height = NULL,
  find_h = TRUE,
  h_max = 1,
  h_min = 0
)
}
\arguments{
\item{distances}{the output object from \code{\link{similarity_to_distance}},
a \code{data.frame} with the first columns called "Site1" and "Site2", and
the other columns being the distance indices or a \code{dist} object}

\item{index}{name of the distance index to use, corresponding to the column
name in \code{distances}. By default, the third column name of
 \code{distances} is used.}

\item{method}{name of the hierarchical classification method, as in
\link[stats:hclust]{stats::hclust()}. Should be one of This should be one of
\code{"ward.D"},
\code{"ward.D2"}, \code{"single"}, \code{"complete"}, \code{"average"}
(= UPGMA), \code{"mcquitty"} (= WPGMA),
\code{"median"} (= WPGMC) or \code{"centroid"} (= UPGMC).}

\item{randomize}{a boolean indicating if the distance matrix should be
randomized, to account for the order of sites in the distance matrix}

\item{n_runs}{number of trials to randomize the distance matrix}

\item{keep_trials}{a boolean indicating if all random trial results
should be stored in the output object (set to FALSE to save space if your
\code{distances} object is large)}

\item{optimal_tree_method}{a character vector indicating how the final tree
should be obtained from all trials. The only option currently is
\code{"best"}, which means the tree with the best cophenetic correlation
coefficient will be chosen.}

\item{n_clust}{an integer or a vector of integers indicating the number of
clusters to be obtained from the hierarchical tree, or the output from
\link{partition_metrics}. Should not be used at the same time as
\code{cut_height}}

\item{cut_height}{a numeric vector indicating the height(s) at which the tree
should be cut. Should not be used at the same time as \code{n_clust}}

\item{find_h}{a boolean indicating if the height of cut should be found for
the requested \code{n_clust}}

\item{h_max}{a numeric indicating the maximum possible tree height for
the chosen \code{index}}

\item{h_min}{a numeric indicating the minimum possible height in the tree for
the chosen \code{index}}
}
\value{
A \code{list} with additional class \code{bioRgeo.hierar.tree}
\itemize{
\item{\code{args}: the input arguments}
\item{\code{dist.matrix}: the distance/beta diversity matrix as a
 \code{dist} object}
\item{\code{trials}: a list containing all randomization trials. Each trial
containes the distance matrix, with site order randomized, the associated
tree and the cophenetic correlation coefficient (spearman) for that tree}
\item{\code{final.tree}: a \code{hclust} object containing the final
hierarchical tree to be used}
\item{\code{final.tree.coph.cor}: the cophenetic correlation coefficient
between the initial distance matrix and \code{final.tree}}
\item{\code{clusters}: a \code{data.frame} containing the clusters}
\item{\code{output_n_clust}: a value or vector specifying the number of
clusters in the output \code{cluster} data.frame}
\item{\code{output_cut_height}: a value or vector specifying, if available,
the height(s) of cut of the tree}
}
}
\description{
This function generates a hierarchical tree from a distance (beta-diversity)
 \code{data.frame},
calculates the cophenetic correlation coefficient,
and can get clusters from the tree if requested by the user. The function
implements randomization of the distance matrix to generate the tree, with
a selection method based on the optimal cophenetic correlation coefficient.
Typically, the distance \code{data.frame} is a \code{bioRgeo.distance} object
obtained by running \code{similarity} and then \code{similarity_to_distance}.
}
\details{
The default method for the hierarchical tree is \code{"average"}, i.e.
UPGMA as it has been recommended as the best method to generate a tree
from beta diversity distances \insertCite{Kreft2010}{bioRgeo}

Clusters can be obtained by two methods:
\itemize{
\item{Specifying a desired number of clusters in \code{n_clust}}
\item{Specifying one or several heights of cut in \code{cut_height}}}

To find an optimal number of clusters, see \code{\link{partition_metrics}}
}
\examples{
comat <- matrix(sample(0:1000, size = 500, replace = TRUE, prob = 1/1:1001), 20, 25)
rownames(comat) <- paste0("Site",1:20)
colnames(comat) <- paste0("Species",1:25)

simil <- similarity(comat, metric = "all")
distances <- similarity_to_distance(simil)

# User-defined number of clusters
#tree1 <- clustering_hierarchical(distances,
#                                 n_clust = 5)
#tree1
#plot(tree1)
#str(tree1)
#tree1$clusters
#
# User-defined height cut
# Only one height
#tree2 <- clustering_hierarchical(distances,
#                                 cut_height = .05)
#tree2
#tree2$clusters
#
# Multiple heights
#tree3 <- clustering_hierarchical(distances,
#                                 cut_height = c(.05, .15, .25))
#tree3
#tree3$clusters # Mind the order of height cuts: from deep to shallow cuts
#
# Recut the tree afterwards
#tree3 <- cut_tree(tree3,
#                  n = 5)
}
\references{
\insertRef{Kreft2010}{bioRgeo}
}
\seealso{
\link{distance_to_similarity}
}
\author{
Pierre Denelle (\email{pierre.denelle@gmail.com}),
Maxime Lenormand (\email{maxime.lenormand@inrae.fr}) and
Boris Leroy (\email{leroy.boris@gmail.com})
}
