% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribution.R
\name{contribution}
\alias{contribution}
\title{Calculate contribution metrics of sites and species}
\usage{
contribution(cluster_object, comat, indices = c("contribution", "Cz"))
}
\arguments{
\item{cluster_object}{a \code{bioregion.clusters} object or a \code{data.frame} or a
list of \code{data.frame} containing multiple partitions. At least two partitions
are required. If a list of \code{data.frame} is provided, they should all have
the same number of rows (i.e., same items in the clustering for all
partitions).}

\item{comat}{a co-occurrence \code{matrix} with sites as rows and species as
columns.}

\item{indices}{a \code{character} specifying the contribution metric to compute.
Available options are \code{contribution}.}
}
\value{
A \code{list} of \code{data.frames} if multiples indices are selected or a single
\code{data.frame} with three columns if one index is selected. Each \code{data.frame}
has three columns: the species, the bioregion, and the contribution
statistics.
}
\description{
Cz metrics
}
\details{
The contribution metric is derived from \insertRef{Lenormand2019}{bioregion}.
Its formula is the following:
\eqn{(n_ij - ((n_i n_j)/n))/(sqrt(((n - n_j)/(n-1)) (1-(n_j/n)) ((n_i n_j)/n)))}

with n the number of sites, n_i the number of sites in which
species i is present, n_j the number of sites belonging to the
bioregion j, n_ij the number of occurrences of species i in sites
belonging to the bioregion j.
}
\examples{
comat <- matrix(sample(1000, 50), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

comat <- matrix(sample(0:1000, size = 500, replace = TRUE, prob = 1/1:1001),
                20, 25)
rownames(comat) <- paste0("Site",1:20)
colnames(comat) <- paste0("Species",1:25)

dissim <- dissimilarity(comat, metric = "Simpson")
clust1 <- nhclu_kmeans(dissim, n_clust = 3, index = "Simpson")

net <- similarity(comat, metric = "Simpson")
com <- netclu_greedy(net)

contribution(cluster_object = clust1, comat = comat,
indices = "contribution")

contribution(cluster_object = com, comat = comat,
indices = "contribution")

}
\seealso{
\link{partition_metrics}
}
\author{
Pierre Denelle (\email{pierre.denelle@gmail.com}),
Boris Leroy (\email{leroy.boris@gmail.com}) and
Maxime Lenormand (\email{maxime.lenormand@inrae.fr})
}
