% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_distance_conversion.R
\name{distance_to_similarity}
\alias{distance_to_similarity}
\title{Convert distance indices to similarity indices}
\usage{
distance_to_similarity(distancedata)
}
\arguments{
\item{distancedata}{the output object from \code{\link{similarity_to_distance}} or a
\code{data.frame} with the first columns called "Site1" and "Site2", and the
other columns being the similarity metrics.}
}
\value{
A \code{data.frame} with additional class \code{bioRgeo.similarity}
}
\description{
This function converts a data.frame of distance indices (beta diversity)
between sites to similarity indices.
}
\note{
The behaviour of this function changes depending on column names. Columns
"Site1" and "Site2" are copied identically. If there are columns called
"a", "b", "c", "A", "B", "C", they will also be copied identically.

If a column is called "Euclidean", its similarity will be calculated based
on the following formula:

\eqn{Euclidean similarity = (1 - Euclidean distance) / Euclidean distance}

Otherwise, all other columns will be transformed into similarity with the
following formula:

\eqn{similarity = 1 - distance}
}
\examples{
comat <- matrix(sample(0:1000, size = 50, replace = TRUE, prob = 1/1:1001), 5, 10)
rownames(comat) <- paste0("Site",1:5)
colnames(comat) <- paste0("Species",1:10)

simil <- spproject(comat, metric = "all")
simil

distances <- similarity_to_distance(simil)
distances

simil <- distance_to_similarity(distances)
simil
}
\seealso{
\link{distance_to_similarity}
}
\author{
Pierre Denelle (\email{pierre.denelle@gmail.com}),
Maxime Lenormand (\email{maxime.lenormand@inrae.fr}) and
Boris Leroy (\email{leroy.boris@gmail.com})
}
