% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evol.distinct.R
\name{evoldistinct.bioregion}
\alias{evoldistinct.bioregion}
\title{A function to calculate evolutionary distinctiveness of bioregions.}
\usage{
evoldistinct.bioregion(P)
}
\arguments{
\item{P}{A distance matrix}
}
\value{
\item{psim}{A site × site phylogenetic beta diversity distance matrix}
}
\description{
The \code{evoldistinct.bioregion} function estimates evolutionary
distinctiveness of each bioregion by computing the mean value of
phylogenetic beta diversity between a focal bioregion and all other
bioregions in the study area.
}
\examples{
example(phylo.beta.pair)
evol.distinct(P)

#2. example
}
\references{
\insertRef{maurin2014savanna}{bioregion}

\insertRef{daru2016novel}{bioregion}

\insertRef{daru2017phylogenetic}{bioregion}

\insertRef{daru2017understanding}{bioregion}
}
\seealso{
\code{\link[picante]{evol.distinct}}
}
\author{
Barnabas H. Daru \email{darunabas@gmail.com} & Klaus Schliep
}
\keyword{cluster}
