% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nclust_tree.R
\name{find_nclust_tree}
\alias{find_nclust_tree}
\title{Find an optimal number of clusters in a hierachical tree based on distances
or beta-diversity indices.}
\usage{
find_nclust_tree(
  tree,
  k_min = 2,
  k_max = "number of sites",
  eval_metric = "pc_distance",
  criterion = "step",
  step_quantile = 0.99,
  step_levels = NULL,
  metric_cutoffs = c(0.5, 0.75, 0.9, 0.95, 0.99, 0.999),
  dist = NULL,
  dist_index = names(dist)[3],
  plot = TRUE,
  disable_progress = FALSE
)
}
\arguments{
\item{tree}{tree a \code{bioRgeo.hierar.tree} or a \code{hclust} object}

\item{k_min}{an integer indicating the minimum number of clusters to be
explored}

\item{k_max}{an integer indicating the maximum number of clusters to be
explored, or "number of sites" to use the number of sites as the maximum}

\item{eval_metric}{metric(s) to be calculated to investigate the effect of
different number of clusters. Note that if you request several evaluation
metrics, they will all be computed, but only the first will be used to
find the optimal number(s) of clusters (order \code{eval_metric} accordingly)}

\item{criterion}{character string indicating the criterion to be used to
identify optimal number(s) of clusters. Available methods currently include
\code{"step"}, \code{"cutoff"}, \code{"elbow"} or \code{"mars"}.}

\item{step_quantile}{if \code{criterion = "step"}, specify here the quantile
of differences between two consecutive k to be used as the cutoff to identify
the most important steps in \code{eval_metric}}

\item{step_levels}{if \code{criterion = "step"}, specify here the number of
steps to keep as cutoffs.}

\item{metric_cutoffs}{if \code{criterion = "cutoff"}, specify here the
cutoffs of \code{eval_metric} at which the number of clusters should be
extracted}

\item{dist}{a \code{dist} object or a \code{bioRgeo.distance} object (output
from \code{\link{similarity_to_distance}}). Necessary if \code{eval_metric}
includes \code{pc_distance}}

\item{dist_index}{a character string indicating the distance (beta-diversity)
index to be used in case \code{dist} is a \code{data.frame} with multiple
distance indices}

\item{plot}{a boolean indicating if a plot of the first \code{eval_metric}
should be drawn with the identified optimal numbers of cutoffs}

\item{disable_progress}{a boolean to enable or disable the progress bar for
the exploration of clusters}
}
\value{
a \code{list} of class \code{bioRgeo.nclust.tree} with three elements:
\itemize{
\item{\code{args}: input arguments
}
\item{\code{evaluation_df}: the data.frame containing \code{eval_metric}
for all explored numbers of clusters
}
\item{\code{optimal_nb_clusters}: a vector containing the optimal number(s)
of cluster(s) according to the first computed \code{eval_metric} and the
chosen \code{criterion}
}}
}
\description{
This function aims at finding an optimal number of clusters in a hierarchical
tree on the basis of the investigation of how an evaluation metric changes as
a function of the number of clusters, and a criterion to find optimal
number(s) of clusters upon this relationship between evaluation metric &
number of clusters.
}
\details{
This function proceeds in three steps. First, the range of clusterisations
between \code{k_min} and \code{k_max} number of clusters are explored on
 \code{tree} by cutting the tree for each number of groups \code{k} between
\code{k_min} and \code{k_max}. Second, an evaluation metric
is calculated for each clusterisation. Third, a criterion is applied on the
evaluation metric to identify the optimal number of clusers.

Note that multiple evaluation metrics can be requested (e.g., for inspection),
but only the first one will be used to apply the criterion for optimal
number of clusters.

\bold{Evaluation metrics:}
\itemize{
\item{\code{pc_distance}: this metric is the method used by
\insertCite{Holt2013}{bioRgeo}. It is a ratio of the between-cluster sum of distances
(beta-diversity)
versus the total sum of distances (beta-diversity) for the full distance
matrix. In
other words, it is calculated on the basis of two elements. First, the total
sum of distances is calculated by summing the entire distance matrix
(\code{dist}). Second, the between-cluster sum of distances is calculated as
follows: for a given number of cluster, the distances are only
summed between clusters, not within clusters. To do that efficiently, all
pairs of sites within the same clusters have their distance set to zero in
the distance matrix, and then the distance matrix is summed. The
\code{pc_distance} ratio is obtained by dividing the between-cluster sum
of distances by the total sum of distances.}
\itemize{\code{pc_endemism}: this metric is the percentage of endemism as
recommended by \insertCite{Kreft2010}{bioRgeo}. To be detailed}
}

\bold{Criteria to find optimal number(s) of clusters}
\itemize{
\item{\code{step}:
This method consists in identifying clusters at the most important
increments, or steps, in the evaluation metric. Therefore, this is relative
to the distribution of increments in the evaluation metric over the tested
\code{k}. Specify \code{step_quantile} as the quantile cutoff above which
increments will be selected as most important (by default, 0.99, i.e. the
top 1\% increments will be selected).
}
\item{\code{cutoffs}:
This method consists in specifying the cutoff value(s) in the evaluation
metric from which the number(s) of clusters should be derived.
}
\item{\code{elbow}:
This method consists in finding an elbow in the evaluation metric curve.
It is based on a fast numerical approximation of the elbow.
The code we use here is based on code written by Alan Tseng in the
KneeArrower R package \url{https://github.com/agentlans/KneeArrower},
and uses code from the
\code{signal} R package written in Octave by Paul Kienzle, modified by
Pascal Dupuis, and translated to R by Tom Short
\url{https://cran.r-project.org/package=signal}}
\item{\code{mars}:
This method consists in fitting a mars model on the evaluation curve, and
using it to identify all cutoffs at which there is no more increase in the
evaluation metric. In other words, this method will find cutoffs with the two
following conditions: (1) the evaluation metric was increasing before the
cutoff and (2) there is no more increase or the increase is slower after the
cutoff. This method uses \link[earth:earth]{earth::earth()}.}
}
}
\note{
Please note that finding the optimal number of clusters is a procedure
which normally requires decisions from the users, and as such can hardly be
fully automatized. Users are strongly advised to read the references
indicated below to look for guidance on how to choose their optimal number(s)
of clusters.
}
\examples{
simil <- spproject(vegemat, metric = "all")
distances <- similarity_to_distance(simil)

# User-defined number of clusters
tree1 <- clustering_hierarchical(distances,
                                 n_clust = 5,
                                 index = "Simpson")
tree1

find_nclust_tree(tree1)

find_nclust_tree(tree1, step_levels = 5)
find_nclust_tree(tree1, eval_metric = "pc_distance", criterion = "elbow")
find_nclust_tree(tree1, eval_metric = "pc_distance", criterion = "mars")
}
\references{
\insertRef{Holt2013}{bioRgeo}

\insertRef{Kreft2010}{bioRgeo}

\insertRef{Langfelder2008}{bioRgeo}
}
\seealso{
\link{clustering_hierarchical}
}
\author{
Pierre Denelle (\email{pierre.denelle@gmail.com}),
Maxime Lenormand (\email{maxime.lenormand@inrae.fr}) and
Boris Leroy (\email{leroy.boris@gmail.com})
}
