% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fishdf}
\alias{fishdf}
\title{Spatial distribution of fish in Europe (data.frame)}
\format{
A \code{data.frame} with 2,703 rows and 3 columns:
\describe{
\item{Site}{Unique site identifier (corresponding to the field ID of fishsf).}
\item{Species}{Unique species identifier.}
\item{Abundance}{Species abundance}
}
}
\usage{
fishdf
}
\description{
A dataset containing the abundance of 195 species in 338 sites.
}
\keyword{datasets}
