% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fishmat}
\alias{fishmat}
\title{Spatial distribution of fish in Europe (co-occurrence matrix)}
\format{
A co-occurrence \code{matrix} with sites as rows and species as
columns. Each element of the matrix
represents the abundance of the species in the site.
}
\usage{
fishmat
}
\description{
A dataset containing the abundance of each of the 195 species in each of
the 338 sites.
}
\keyword{datasets}
