% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishnet.R
\name{fishnet}
\alias{fishnet}
\title{Creating fishnet of polygons}
\usage{
fishnet(shp, res = 0.5)
}
\arguments{
\item{files}{A community matrix}
}
\value{
\item{psim}{A site × site phylogenetic beta diversity distance matrix}
}
\description{
The \code{fishnet} function enables the process in which species occurrence
data and ecological and environmental variables are modeled together to
predict species ranges.
}
\examples{
library(raster)
file <- paste0(system.file(package="bioregion"), "/inst/nigeria/nigeria.shp")
d <- shapefile(file)
d1 <- fishnet(d, res = 0.75)
}
\references{
\insertRef{Philips2006}{bioregion}
}
\keyword{bioregion}
