% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Clusters.R
\name{getClusters}
\alias{getClusters}
\title{Gets the descendant nodes of a phylogeny at a given time depth.}
\usage{
getClusters(tree, cut = 2, ...)
}
\arguments{
\item{tree}{is a dated phylogenetic tree with branch lengths stored as a 
phylo object (as in the \code{ape} package).}
}
\value{
\item{}{A list of descendants}
}
\description{
\code{getClusters} returns the tips that descend from a given node or time depth on a dated phylogenetic tree.
}
\examples{
require(phylobase)
geo <- get(data(geospiza))
tree <- geo$tree
plot(tree)
axisPhylo(side = 1)
getClusters(tree, .3)
}
\references{
\insertRef{Schliep2010}{bioregion}
}
\keyword{bioregion}
