% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_prop.R
\name{label_prop}
\alias{label_prop}
\title{Finding communities based on propagating labels}
\usage{
label_prop(net, weight = TRUE, directed = FALSE)
}
\arguments{
\item{net}{a two- or three-column \code{data.frame} representing a network with the two first columns
as (un)directed links between pair of nodes and an optional third column indicating the weight of the link}

\item{weight}{a boolean indicating if the weights should be considered if there is a third column}
}
\value{
A \code{data.frame} providing one community by node.
}
\description{
This function finds communities in a (un)weighted undirected network based on propagating labels.
}
\details{
This function is based on propagating labels \insertCite{Raghavan2007}{bioRgeo}
as implemented in the \href{https://cran.r-project.org/web/packages/igraph/index.html}{igraph} package
(\link[igraph]{cluster_label_prop}).
}
\examples{
comat=matrix(sample(1000,50),5,10)
rownames(comat)=paste0("Site",1:5)
colnames(comat)=paste0("Species",1:10)

net=spproject(comat,metric="Simpson")
com=label_prop(net)
}
\references{
\insertRef{Raghavan2007}{bioRgeo}
}
\seealso{
\link{infomap}, \link{oslom}
}
\author{
Pierre Denelle (\email{pierre.denelle@gmail.com}),
Maxime Lenormand (\email{maxime.lenormand@inrae.fr}) and
Boris Leroy (\email{leroy.boris@gmail.com})
}
