% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_bioregions.R
\name{map_bioregions}
\alias{map_bioregions}
\title{Create a map of bioregions}
\usage{
map_bioregions(
  bioregionalization,
  map,
  partition_index = NULL,
  map_as_output = FALSE,
  plot = TRUE,
  clusters = NULL,
  geometry = NULL,
  write_clusters = NULL,
  ...
)
}
\arguments{
\item{bioregionalization}{A \code{bioregion.clusters} object.}

\item{map}{A spatial object that can be handled by \code{sf} or \code{terra}.
The first attribute or layer should correspond to the sites' ID
(see Details).}

\item{partition_index}{An \code{integer}, \code{character}, or \code{NULL} specifying
which \code{bioregionalization}'s partition(s) to plot. By default (\code{NULL}), all
partitions are plotted. If an \code{integer} or vector of \code{integers} is provided,
partition(s) are selected by column number(s) in the \code{bioregionalization}
data.frame. If a \code{character} or vector of \code{characters}, partition(s) are
selected by name(s) matching column names in \code{bioregionalization}.}

\item{map_as_output}{A \code{boolean} indicating if the \code{sf} \code{data.frame} object
used for the plot should be returned.}

\item{plot}{A \code{boolean} indicating if the plot should be drawn.}

\item{clusters}{Deprecated. Use \code{bioregionalization} instead. The former
\code{bioregionalization} has been replaced by \code{partition_index}.}

\item{geometry}{Deprecated. Use \code{map} instead.}

\item{write_clusters}{Deprecated. Use \code{map_as_output} instead.}

\item{...}{Further arguments to be passed to \code{sf::plot()}.}
}
\value{
One or several maps of bioregions if \code{plot = TRUE} and the
\code{sf} \code{data.frame} object used for the plot if \code{map_as_output = TRUE}.
}
\description{
This plot function can be used to visualize bioregions based on a
\code{bioregion.clusters} object combined with a spatial object (\code{sf} or \code{terra}).
}
\details{
The site IDs in \code{bioregionalization} and \code{map} should correspond. They must
have the same type (i.e., \code{character} if \code{bioregionalization} is a
\code{bioregion.clusters} object), and the sites in \code{bioregionalization} should be
included among the sites in \code{map}. If \code{map} is an \code{sf} or a \code{SpatVector}
(\code{terra}) object, it should contain an attribute table with the IDs in the
first column. If \code{map} is a \code{SpatRaster} (\code{terra}) object, it should contain
the IDs in the first layer.

If the \code{bioregionalization} object contains both types of nodes (sites and
species), only site  will be mapped. The function automatically filters to
site nodes using the \code{node_type} attribute.

\strong{Colors}: If the \code{bioregionalization} object contains colors (added via
\code{bioregion_colors()}), these colors will be automatically used for plotting.
Otherwise, the default \code{sf} color scheme will be applied.
}
\examples{
data(fishmat)
data(fishsf)

net <- similarity(fishmat, metric = "Simpson")
clu <- netclu_greedy(net)
mapclu <- map_bioregions(clu, 
                         map = fishsf, 
                         map_as_output = TRUE, 
                         plot = FALSE)

# With colors
clu_colored <- bioregion_colors(clu)
mapclu <- map_bioregions(clu_colored, 
                         map = fishsf, 
                         map_as_output = TRUE, 
                         plot = FALSE)
           
}
\seealso{
For more details illustrated with a practical example,
see the vignette:
\url{https://biorgeo.github.io/bioregion/articles/a5_1_visualization.html}.

Associated functions:
\link{bioregion_colors}
}
\author{
Maxime Lenormand (\email{maxime.lenormand@inrae.fr}) \cr
Boris Leroy (\email{leroy.boris@gmail.com}) \cr
Pierre Denelle (\email{pierre.denelle@gmail.com})
}
