% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netclu_greedy.R
\name{netclu_greedy}
\alias{netclu_greedy}
\title{Community structure detection via greedy optimization of modularity}
\usage{
netclu_greedy(net, weight = TRUE)
}
\arguments{
\item{net}{a two- or three-column \code{data.frame} representing a network with the two first columns
as undirected links between pair of nodes and an optional third column indicating the weight of the link}

\item{weight}{a boolean indicating if the weights should be considered if there is a third column}
}
\value{
A \code{data.frame} providing one community by node.
}
\description{
This function finds communities in a (un)weighted undirected network via greedy optimization of modularity.
}
\details{
This function is based on the fast greedy modularity optimization algorithm \insertCite{Clauset2004}{bioRgeo}
as implemented in the \href{https://cran.r-project.org/web/packages/igraph/index.html}{igraph} package
(\link[igraph]{cluster_fast_greedy}).
}
\examples{
comat=matrix(sample(1000,50),5,10)
rownames(comat)=paste0("Site",1:5)
colnames(comat)=paste0("Species",1:10)

net=similarity(comat,metric="Simpson")
com=netclu_greedy(net)
}
\references{
\insertRef{Clauset2004}{bioRgeo}
}
\seealso{
\link{netclu_infomap}, \link{netclu_oslom}
}
\author{
Pierre Denelle (\email{pierre.denelle@gmail.com}),
Maxime Lenormand (\email{maxime.lenormand@inrae.fr}) and
Boris Leroy (\email{leroy.boris@gmail.com})
}
