% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhclu_pam.R
\name{nhclu_pam}
\alias{nhclu_pam}
\title{Non hierarchical clustering: partitioning around medoids}
\usage{
nhclu_pam(
  distances,
  index = names(distances)[3],
  n_clust = NULL,
  nstart = 10,
  variant = "faster",
  cluster_only = FALSE,
  ...
)
}
\arguments{
\item{distances}{the output object from \code{\link{similarity_to_distance}},
a \code{data.frame} with the first columns called "Site1" and "Site2", and
the other columns being the distance indices or a \code{dist} object}

\item{index}{a \code{character} string providing the name of the distance
index to use, corresponding to the column
name in \code{distances}. By default, the third column name of
 \code{distances} is used.}

\item{n_clust}{an \code{integer} specifying the requested number of cluster for pam
and kmeans.}

\item{nstart}{an \code{integer} specifying the number of random “starts”
for the pam algorithm. By default, 1 (for the \code{"faster"} variant)}

\item{variant}{a \code{character} string specifying the variant of pam
to use, by default "faster". See \link[cluster:pam]{cluster::pam()} for
more details}

\item{cluster_only}{a \code{boolean} specifying if only the clustering
should be returned from the \link[cluster:pam]{cluster::pam()} function
(more efficient)}

\item{...}{you can add here further arguments to be passed to \code{pam()}
(see \link[cluster:pam]{cluster::pam()})}
}
\value{
A \code{list} of class \code{bioRgeo.clusters} with five slots:
\enumerate{
\item{\bold{name}: \code{character string} containing the name of the algorihtm}
\item{\bold{args}: \code{list} of input arguments as provided by the user}
\item{\bold{inputs}: \code{list} of characteristics of the input dataset}
\item{\bold{algorithm}: \code{list} of all objects associated with the
 clustering procedure, such as original cluster objects}
\item{\bold{clusters}: \code{data.frame} containing the clustering results}}
}
\description{
This function performs non hierarchical
clustering on the basis of distances with partioning around medoids.
}
\details{
This method partitions data into
 the chosen number of cluster on the basis of the input distance matrix.
 It is more robust than k-means because it minimizes the sum of distances
 between cluster centres and points assigned to the cluster -
 whereas the k-means approach minimizes the sum of squared euclidean
 distances (thus k-means cannot be applied directly on the input distance
 matrix if the distances are not euclidean).
}
\examples{
simil <- similarity(vegemat, metric = "all")
distances <- similarity_to_distance(simil)

clust1 <- nhclu_pam(distances,
    n_clust = 2:10,
    index = "Simpson")
clust2 <- nhclu_pam(distances,
    n_clust = 2:25,
    index = "Simpson")
partition_metrics(clust2,
                  distances = distances,
                  eval_metric = "pc_distance",
                  partition_optimisation = TRUE)
partition_metrics(clust2,
                  sp_site_table = vegemat,
                  eval_metric = "avg_endemism",
                  partition_optimisation = TRUE)
}
