% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oslom.R
\name{oslom}
\alias{oslom}
\title{OSLOM community finding}
\usage{
oslom(
  net,
  weight = TRUE,
  reassign = "no",
  r = 10,
  hr = 50,
  seed = 1,
  t = 0.1,
  cp = 0.5,
  directed = FALSE,
  delete_temp = TRUE,
  path_temp = "oslom_temp",
  binpath = NULL
)
}
\arguments{
\item{net}{a two- or three-column \code{data.frame} representing a network with the two first columns
as (un)directed links between pair of nodes and an optional third column indicating the weight of the link}

\item{weight}{a boolean indicating if the weights should be considered if there is a third column}

\item{reassign}{a string indicating if the nodes belonging to several community should be reassign
and what method should be used (see Details)}

\item{r}{the number of runs for the first hierarchical level (10 by default)}

\item{hr}{the number of runs for the higher hierarchical level (50 by default, 0 if you are not interested
in hierarchies)}

\item{seed}{for the random number generator}

\item{t}{the p−value, the default value is 0.10, increase this value you to get more modules}

\item{cp}{kind of resolution parameter used to decide between taking some modules or their union (default value is
0.5, bigger value leads to bigger clusters)}

\item{directed}{a boolean indicating if the network is directed (from column 1 to column 2)}

\item{delete_temp}{a boolean indicating if the temporary folder should be removed (see Details)}

\item{path_temp}{a string indicating the path to the temporary folder (see Details)}

\item{binpath}{a string indicating the path to the bin folder (see \link{install_binaries} and Details)}
}
\value{
A \code{data.frame} providing one or several modules (according to the chosen option(s)) for each node. If
\code{reassign = simil} or \code{reassign = random} only one column by hierarchical level will be provided. If
\code{reassign = no} the number of columns will depend on the number of overlapping modules (up to three). A value of 0
indicates that no module were found (i.e. non-overlapping nodes).
}
\description{
This function finds communities in a (un)weighted (un)directed network based on the OSLOM algorithm
(\url{http://oslom.org/}, version 2.4).
}
\details{
OSLOM is a network community detection algorithm proposed in \insertCite{Lancichinetti2011}{bioRgeo} that
finds statistically significant (overlapping) communities in (un)weighted and (un)directed networks. Since a node
may belong to several communities we propose two methods to reassign the 'overlapping' nodes randomly
\code{reassign = 'random'} or based on the closest candidate community \code{reassign = 'simil'}
(only for weighted networks, in this case the closest candidate community is determined with the average similarity).
By default \code{reassign = 'no'} and all the information is provided.

This function is based on the 2.4 C++ version of OSLOM (\url{http://www.oslom.org/software.htm}).
This function needs executable files to run. They can be installed with \link{install_binaries}. If you set the path to
the folder that will host the bin folder  manually while running \link{install_binaries} please make sure to set \code{binpath}
accordingly.

The C++ version of OSLOM generates temporary folders and/or files that are stored in the \code{path_temp} folder
(folder "oslom_temp" in the working directory by default). This temporary folder is removed by default
(\code{delete_temp = TRUE}).
}
\examples{
comat=matrix(sample(1000,50),5,10)
rownames(comat)=paste0("Site",1:5)
colnames(comat)=paste0("Species",1:10)

net=similarity(comat,metric="Simpson")
#com=oslom(net) # run bin() to use this function
}
\references{
\insertRef{Lancichinetti2011}{bioRgeo}
}
\seealso{
\link{infomap}, \link{louvain}
}
\author{
Pierre Denelle (\email{pierre.denelle@gmail.com}),
Maxime Lenormand (\email{maxime.lenormand@inrae.fr}) and
Boris Leroy (\email{leroy.boris@gmail.com})
}
