% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster2comm.R
\name{raster2comm}
\alias{raster2comm}
\title{Converts raster to community matrix}
\usage{
raster2comm(files)
}
\arguments{
\item{files}{File names}
}
\description{
\code{raster2comm} converts a set of raster files from a species distribution
modeling into a community matrix.
}
\examples{
fdir <- system.file("Aloes", package="bioregion")
files <- file.path(fdir, dir(fdir))
res <- raster2comm(files)

}
\keyword{bioregion}
