% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm.R
\name{sdm}
\alias{sdm}
\title{Global scale species distribution modeling}
\usage{
sdm(x)
}
\arguments{
\item{files}{A community matrix}
}
\value{
\item{psim}{A site × site phylogenetic beta diversity distance matrix}
}
\description{
The \code{sdm} function enables the process in which species occurrence
data and ecological and environmental variables are modeled together to
predict species ranges.
}
\examples{
fdir <- system.file("Aloes", package="bioregion")
files <- file.path(fdir, dir(fdir))
res <- raster2comm(files)
}
\references{
\insertRef{Philips2006}{bioregion}
}
\keyword{bioregion}
