% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{similarity}
\alias{similarity}
\title{Compute similarity metrics between sites based on species composition}
\usage{
similarity(comat, metric = "Simpson", formula = NULL, method = "prodmat")
}
\arguments{
\item{comat}{a co-occurrence \code{matrix} with sites as rows and species as
columns.}

\item{metric}{a \code{character} vector or a single \code{character} string indicating
which metrics to chose (see Details). Available options are \code{"abc"}, \code{"ABC"},
\code{"Jaccard"}, \code{"Jaccardturn"}, \code{"Sorensen"}, \code{"Simpson"}, \code{"Bray"},
\code{"Brayturn"} or \code{"Euclidean"}. If \code{"all"} is specified, then all metrics
will be calculated. Can be set to \code{NULL} if \code{formula} is used.}

\item{formula}{a \code{character} vector or a single \code{character} string with your
own formula(s) based on the \code{a}, \code{b}, \code{c}, \code{A}, \code{B}, and \code{C} quantities
(see Details). \code{formula} is set to \code{NULL} by default.}

\item{method}{a \code{character} string indicating what method should be used to
compute \code{abc} (see Details). \code{method = "prodmat"} by default is more
efficient but can be greedy in memory and \code{method = "loops"} is less
efficient but less greedy in memory.}
}
\value{
A \code{data.frame} with additional class
\code{bioregion.pairwise.metric}, providing one or several similarity
metric(s) between each pair of sites. The two first columns represent each
pair of sites. One column per similarity metric provided in \code{metric} and
\code{formula} except for the metric \code{abc} and \code{ABC} that are
stored in three columns (one for each letter).
}
\description{
This function creates a \code{data.frame} where each row provides one or
several similarity metric(s) between each pair of sites from a co-occurrence
\code{matrix} with sites as rows and species as columns.
}
\details{
With \code{a} the number of species shared by a pair of sites, \code{b}
species only present in the first site and \code{c} species only present in
the second site.

Jaccard = 1 - (b + c) / (a + b + c)

Jaccardturn = 1 - 2min(b, c) / (a + 2min(b, c)) (Baselga, 2012)

Sorensen = 1 - (b + c) / (2a + b + c)

Simpson = 1 - min(b, c) / (a + min(b, c))

If abundances data are available, Bray-Curtis and its turnover component can
also be computed with the following equation:

Bray = 1 - (B + C) / (2A + B + C)

Brayturn = 1 - min(B, C) / (A + min(B, C)) (Baselga, 2013)

with A the sum of the lesser values for common species shared by a pair of
sites.
B and C are the total number of specimens counted at both sites minus A.

\code{formula} can be used to compute customized metrics with the terms
\code{a}, \code{b}, \code{c}, \code{A}, \code{B}, and \code{C}. For example
\code{formula = c("1 - pmin(b,c) / (a + pmin(b,c))", "1 - (B + C) / (2*A + B + C)")}
will compute the Simpson and Bray-Curtis similarity metrics, respectively.
Note that pmin is used in the Simpson formula because a, b, c, A, B and C
are \code{numeric} vectors.

Euclidean computes the Euclidean similarity between each pair of site
following this equation:

Euclidean = 1 / (1 + d_ij)

Where d_ij is the Euclidean distance between site i and
site j in terms of species composition.
}
\examples{
comat <- matrix(sample(0:1000, size = 50, replace = TRUE,
prob = 1 / 1:1001), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

sim <- similarity(comat, metric = c("abc", "ABC", "Simpson", "Brayturn"))

sim <- similarity(comat, metric = "all",
formula = "1 - (b + c) / (a + b + c)")

}
\references{
Baselga A (2012) The Relationship between Species Replacement,
Dissimilarity Derived from Nestedness, and Nestedness.
\emph{Global Ecology and Biogeography}, 21(12), 1223--1232.

Baselga A (2013) Separating the two components of abundance-based
dissimilarity: balanced changes in abundance vs. abundance gradients.
\emph{Methods in Ecology and Evolution}, 4(6), 552--557.
}
\seealso{
\link{dissimilarity} \link{dissimilarity_to_similarity}
\link{similarity_to_dissimilarity}
}
\author{
Maxime Lenormand (\email{maxime.lenormand@inrae.fr}) \cr
Pierre Denelle (\email{pierre.denelle@gmail.com}) \cr
Boris Leroy (\email{leroy.boris@gmail.com})
}
