% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_distance_conversion.R
\name{similarityToDistance}
\alias{similarityToDistance}
\title{Convert similarity indices to distance indices}
\usage{
similarityToDistance(similaritydata)
}
\arguments{
\item{similaritydata}{the output object from \code{\link{spproject}} or a
\code{data.frame} with the first columns called "Site1" and "Site2", and the
other columns being the similarity metrics.}
}
\value{
A \code{data.frame} with additional class \code{bioRgeo.distance}
}
\description{
This function converts a data.frame of similarity indices between sites to
 distance indices (= beta diversity).
}
\note{
The behaviour of this function changes depending on column names. Columns
"Site1" and "Site2" are copied identically. If there are columns called
"a", "b", "c", "A", "B", "C", they will also be copied identically.

If a column is called "Euclidean", its distance will be calculated based
on the following formula:

\eqn{Euclidean distance = (1 - Euclidean similarity) / Euclidean similarity}

Otherwise, all other columns will be transformed into distances with the
following formula:

\eqn{distance = 1 - similarity}
}
\examples{
comat <- matrix(sample(0:1000, size = 50, replace = TRUE, prob = 1/1:1001), 5, 10)
rownames(comat) <- paste0("Site",1:5)
colnames(comat) <- paste0("Species",1:10)

simil <- spproject(comat, metric = "all")
simil

distances <- similarityToDistance(simil)
distances
}
\seealso{
\link{distanceToSimilarity}
}
\author{
Pierre Denelle (\email{pierre.denelle@gmail.com}),
Maxime Lenormand (\email{maxime.lenormand@inrae.fr}) and
Boris Leroy (\email{leroy.boris@gmail.com})
}
