% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_species_subset.R
\name{site_species_subset}
\alias{site_species_subset}
\title{Extract a subset of sites or species from a \code{bioregion.clusters} object}
\usage{
site_species_subset(bioregionalization, node_type = "site", clusters = NULL)
}
\arguments{
\item{bioregionalization}{An object of class \code{bioregion.clusters}.}

\item{node_type}{A \code{character} string indicating the type of nodes to
extract. Possible values are \code{"site"} or \code{"species"}. The default is
\code{"site"}.}

\item{clusters}{Deprecated. Use \code{bioregionalization} instead.}
}
\value{
An object of class \code{bioregion.clusters} containing only the specified
node type (sites or species).
}
\description{
This function extracts a subset of nodes based on their type (\code{"site"} or
\code{"species"}) from a \code{bioregion.clusters} object, which contains both types of
nodes (sites and species).
}
\note{
Some \code{bioregion.clusters} objects may contain both types of nodes (sites and
species). This information is available in the \verb{$inputs$node_type} slot.

This function allows you to extract a specific type of node
(either sites or species) from any \code{bioregion.clusters} object that
includes both.
}
\examples{
net <- data.frame(
  Site = c(rep("A", 2), rep("B", 3), rep("C", 2)),
  Species = c("a", "b", "a", "c", "d", "b", "d"),
  Weight = c(10, 100, 1, 20, 50, 10, 20)
)

clu <- netclu_louvain(net, lang = "igraph", bipartite = TRUE)

clu_sites <- site_species_subset(clu, node_type = "site")

}
\author{
Maxime Lenormand (\email{maxime.lenormand@inrae.fr}) \cr
Pierre Denelle (\email{pierre.denelle@gmail.com}) \cr
Boris Leroy (\email{leroy.boris@gmail.com})
}
