% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spproject.R
\name{spproject}
\alias{spproject}
\title{Compute similarity metrics between sites based on species composition}
\usage{
spproject(comat, metric = "Simpson", method = "prodmat")
}
\arguments{
\item{comat}{a co-occurence \code{matrix} with sites as rows and species as columns}

\item{metric}{a vector of character(s) indicating which similarity
metric(s) to chose (see Details). If \code{"all"} is specified, then all
metrics will be calculated.}

\item{method}{a string indicating what method should be used to compute \code{abc} (see Details),
\code{method = "prodmat"} by default is more efficient but can greedy in memory and \code{method="loops"} is less efficient
but less greedy in memory}
}
\value{
A \code{data.frame} providing one or several similarity
metric(s) between each pair of sites. The two first columns represents each pair of sites.
One column per similarity metric provided in \code{metric} except for the metric \emph{abc} and \emph{ABC} that
are stored in three columns (one for each letter).
}
\description{
This function creates a \code{data.frame} where each row provides one or several similarity
metric(s) between each pair of sites from a co-occurence \code{matrix} with sites as rows and species as columns.
}
\details{
With \code{a} the number of species shared by a pair of sites, \code{b} species only present in the first site
and \code{c} species only present in the second site.

\eqn{Jaccard = 1 - (b + c) / (a + b + c)}

\eqn{Jaccardturn = 1 - 2min(b, c) / (a + 2min(b, c))} \insertCite{Baselga2012}{bioRgeo}

\eqn{Sorensen = 1 - (b + c) / (2a + b + c)}

\eqn{Simpson = 1 - min(b, c) / (a + min(b, c))}

If abundances data are available, Bray-Curtis and its turnover component can also be computed with the
following equation:

\eqn{Bray = 1 - (B + C) / (2A + B + C)}

\eqn{Brayturn = 1 - min(B, C)/(A + min(B, C))} \insertCite{Baselga2013}{bioRgeo}

with A the sum of the lesser values for common species shared by a pair of sites.
B and C are the total number of specimens counted at both sites minus A.

Euclidean computes the Euclidean similarity between each pair of site following this equation:

\eqn{Euclidean = 1 / (1 + dij)}

Where dij is the Euclidean distance between site i and site j in terms of species composition.
}
\examples{
comat <- matrix(sample(0:1000, size = 50, replace = TRUE, prob = 1/1:1001), 5, 10)
rownames(comat) <- paste0("Site",1:5)
colnames(comat) <- paste0("Species",1:10)

simil <- spproject(comat, metric = c("abc", "ABC", "Simpson", "Brayturn"))
simil

simil <- spproject(comat, metric = "all")
simil

}
\references{
\insertRef{Baselga2012}{bioRgeo}
\insertRef{Baselga2013}{bioRgeo}
}
\seealso{
\link{distance_to_similarity}
}
\author{
Pierre Denelle (\email{pierre.denelle@gmail.com}),
Maxime Lenormand (\email{maxime.lenormand@inrae.fr}) and
Boris Leroy (\email{leroy.boris@gmail.com})
}
