% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vegemat}
\alias{vegemat}
\title{Spatial distribution of Mediterrean Vegetation}
\format{
A
\describe{
  \item{vegedf}{  }
  \item{vegemat}{  }
  \item{vegesp}{  }
}
}
\source{
<http://>
}
\usage{
vegemat
}
\description{
A dataset containing the
}
\keyword{datasets}
