% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walktrap.R
\name{walktrap}
\alias{walktrap}
\title{Community structure detection via short random walks}
\usage{
walktrap(net, weight = TRUE, steps = 4)
}
\arguments{
\item{net}{a two- or three-column \code{data.frame} representing a network with the two first columns
as undirected links between pair of nodes and an optional third column indicating the weight of the link}

\item{weight}{a boolean indicating if the weights should be considered if there is a third column}

\item{steps}{the length of the random walks to perform}
}
\value{
A \code{data.frame} providing one community by node.
}
\description{
This function finds communities in a (un)weighted undirected network via short random walks.
}
\details{
This function is based on random walks \insertCite{Pons2005}{bioRgeo}
as implemented in the \href{https://cran.r-project.org/web/packages/igraph/index.html}{igraph} package
(\link[igraph]{cluster_walktrap}).
}
\examples{
comat=matrix(sample(1000,50),5,10)
rownames(comat)=paste0("Site",1:5)
colnames(comat)=paste0("Species",1:10)

net=similarity(comat,metric="Simpson")
com=walktrap(net)
}
\references{
\insertRef{Pons2005}{bioRgeo}
}
\seealso{
\link{infomap}, \link{oslom}
}
\author{
Pierre Denelle (\email{pierre.denelle@gmail.com}),
Maxime Lenormand (\email{maxime.lenormand@inrae.fr}) and
Boris Leroy (\email{leroy.boris@gmail.com})
}
