% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedEndemism.R
\name{weighted.endemism}
\alias{weighted.endemism}
\title{Read in sparse community matrices}
\usage{
weighted.endemism(x)
}
\arguments{
\item{files}{A community matrix}
}
\value{
\item{psim}{A site × site phylogenetic beta diversity distance matrix}
}
\description{
\code{weighted.endemism} is species richness inversely weighted
by species ranges.
}
\examples{
fdir <- system.file("Aloes", package="bioregion")
files <- file.path(fdir, dir(fdir))
res <- raster2comm(files)
}
\references{
\insertRef{Crispetal2001}{bioregion}

\insertRef{LaffanCrisp2003}{bioregion}
}
\keyword{bioregion}
