% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioregionalization_metrics.R
\name{bioregionalization_metrics}
\alias{bioregionalization_metrics}
\title{Calculate metrics for one or several partitions}
\usage{
bioregionalization_metrics(
  cluster_object,
  dissimilarity = NULL,
  dissimilarity_index = NULL,
  net = NULL,
  site_col = 1,
  species_col = 2,
  eval_metric = c("pc_distance", "anosim", "avg_endemism", "tot_endemism")
)
}
\arguments{
\item{cluster_object}{a \code{bioregion.clusters} object.}

\item{dissimilarity}{a \code{dist} object or a \code{bioregion.pairwise.metric}
object (output from \code{\link[=similarity_to_dissimilarity]{similarity_to_dissimilarity()}}). Necessary if
\code{eval_metric} includes \code{pc_distance} and \code{tree} is not a
\code{bioregion.hierar.tree} object}

\item{dissimilarity_index}{a \code{character} string indicating the dissimilarity
(beta-diversity) index to be used in case \code{dist} is a \code{data.frame} with
multiple dissimilarity indices.}

\item{net}{the species-site network (i.e., bipartite network). Should be
provided as \verb{data.frame± if }eval_metric\code{includes}"avg_endemism"\code{or }"tot_endemism"`.}

\item{site_col}{name or number for the column of site nodes (i.e. primary
nodes). Should be provided if \code{eval_metric} includes \code{"avg_endemism"} or
\code{"tot_endemism"}.}

\item{species_col}{name or number for the column of species nodes (i.e.
feature nodes). Should be provided if \code{eval_metric} includes
\code{"avg_endemism"} or \code{"tot_endemism"}.}

\item{eval_metric}{a \code{character} vector or a single \code{character} string
indicating metric(s) to be calculated to investigate the effect of different
number of clusters. Available options are \code{"pc_distance"}, \code{"anosim"},
\code{"avg_endemism"} or \code{"tot_endemism"}.}
}
\value{
a \code{list} of class \code{bioregion.partition.metrics} with two to three elements:
\itemize{
\item{\code{args}: input arguments
}
\item{\code{evaluation_df}: the data.frame containing \code{eval_metric}
for all explored numbers of clusters
}
\item{\code{endemism_results}: if endemism calculations were requested, a list
with the endemism results for each partition
}
}
}
\description{
This function aims at calculating metrics for one or several partitions,
usually on outputs from \code{netclu_}, \code{hclu_} or \code{nhclu_} functions. Metrics
may require the users to provide either a similarity or dissimilarity
matrix, or to provide the initial species-site table.
}
\details{
\strong{Evaluation metrics:}

\itemize{

\item{\code{pc_distance}: this metric is the method used by
Holt et al. (2013). It is a ratio of the between-cluster sum of
dissimilarity (beta-diversity) versus the total sum of dissimilarity
(beta-diversity) for the full dissimilarity matrix. In other words, it is
calculated on the basis of two elements. First, the total sum of
dissimilarity is calculated by summing the entire dissimilarity matrix
(\code{dist}). Second, the between-cluster sum of dissimilarity is calculated as
follows: for a given number of cluster, the dissimilarity is only summed
between clusters, not within clusters. To do that efficiently, all pairs of
sites within the same clusters have their dissimilarity set to zero in
the dissimilarity matrix, and then the dissimilarity matrix is summed. The
\code{pc_distance} ratio is obtained by dividing the between-cluster sum of
dissimilarity by the total sum of dissimilarity.}

\item{\code{anosim}: This metric is the statistic used in Analysis of
Similarities, as suggested in Castro-Insua et al. (2018) (see
\link[vegan:anosim]{vegan::anosim()}). It compares the between-cluster
dissimilarities to the within-cluster dissimilarities. It is based based on
the difference of mean ranks between groups and within groups with the
following formula:
R = (r_B - r_W)/(N (N-1) / 4),
where r_B and r_W are the average ranks
between and within clusters respectively, and N is the total
number of sites.
Note that the function does not estimate the significance here, it only
computes the statistic - for significance testing see
\link[vegan:anosim]{vegan::anosim()}}.

\item{\code{avg_endemism}: this metric is the average percentage of
endemism in clusters as
recommended by Kreft & Jetz (2010). Calculated as follows:
End_mean = sum_i (E_i / S_i)/K
where E_i is the number of endemic species in cluster i, S_i is the number of
species in cluster i, and K the maximum number of clusters.
}

\item{\code{tot_endemism}: this metric is the total endemism across all clusters,
as recommended by Kreft & Jetz (2010). Calculated as follows:
End_tot = E \ C
where E is total the number of endemics (i.e., species found in
only one cluster) and C is the number of non-endemic species.
}
}
}
\examples{
comat <- matrix(sample(0:1000, size = 500, replace = TRUE, prob = 1/1:1001),
20, 25)
rownames(comat) <- paste0("Site",1:20)
colnames(comat) <- paste0("Species",1:25)

comnet <- mat_to_net(comat)

dissim <- dissimilarity(comat, metric = "all")

# User-defined number of clusters
tree1 <- hclu_hierarclust(dissim, 
                          n_clust = 2:20, index = "Simpson")
tree1

a <- bioregionalization_metrics(tree1, dissimilarity = dissim, net = comnet,
                       site_col = "Node1", species_col = "Node2",
                       eval_metric = c("tot_endemism", "avg_endemism",
                                      "pc_distance", "anosim"))
a

}
\references{
Castro-Insua A, Gómez-Rodríguez C & Baselga A (2018) Dissimilarity measures
affected by richness differences yield biased delimitations of biogeographic
realms. \emph{Nature Communications}, 9(1), 9-11.

Holt BG, Lessard J, Borregaard MK, Fritz SA, Araújo MB, Dimitrov D, Fabre P,
Graham CH, Graves GR, Jønsson Ka, Nogués-Bravo D, Wang Z, Whittaker RJ,
Fjeldså J & Rahbek C (2013) An update of Wallace's zoogeographic regions of
the world. \emph{Science}, 339(6115), 74-78.

Kreft H & Jetz W (2010) A framework for delineating biogeographical regions
based on species distributions. \emph{Journal of Biogeography}, 37, 2029-2053.
}
\seealso{
\link{compare_bioregionalizations}
}
\author{
Boris Leroy (\email{leroy.boris@gmail.com}) \cr
Maxime Lenormand (\email{maxime.lenormand@inrae.fr}) \cr
Pierre Denelle (\email{pierre.denelle@gmail.com})
}
