% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth.R
\name{choropleth}
\alias{choropleth}
\title{Binning shapefile polygons based on slot values}
\usage{
choropleth(x, values, k = 10, ...)
}
\arguments{
\item{x}{An object of the class SpatialPolygonsDataFrame.}

\item{values}{The column in the SpatialPolygonsDataFrame for which to discretize 
the values of the quantity.}

\item{k}{Numeric, the desired number of bins to discretize.}

\item{\dots}{Further arguments passed to or from other methods.}
}
\value{
\item{choropleth} {returns a SpatialPolygonsDataFrame with a column with the 
discretized values}
}
\description{
The \code{choropleth} function discretizes the values of a quantity based 
on their quantiles.
}
\examples{
x <- c("raster", "colorRamps")
lapply(x, library, character.only = TRUE)
s <- readRDS(system.file("nigeria/SR_Naija.rds", package= "bioregion"))
k=10
COLOUR <- blue2green2red(k)
y = choropleth(s, values=s$SR, k)
 
## Not run: 
plot(y, col=COLOUR[y$values], border = NA)
 
## End(Not run)

}
\keyword{cluster}
