% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_dissimilarity_conversion.R
\name{dissimilarity_to_similarity}
\alias{dissimilarity_to_similarity}
\title{Convert dissimilarity indices to similarity indices}
\usage{
dissimilarity_to_similarity(dissimilaritydata)
}
\arguments{
\item{dissimilaritydata}{the output object from
\code{\link[=similarity_to_dissimilarity]{similarity_to_dissimilarity()}} or a \code{data.frame} with the first columns
called "Site1" and "Site2", and the other columns being the similarity
metrics.}
}
\value{
A \code{data.frame} with additional class \code{bioRgeo.pairwise.metric},
providing one or several dissimilarity metric(s) between each pair of sites.
The two first columns represent each pair of sites, and the other column
represent dissimilarity metrics. Columns with names "a", "b", "c", "A", "B"
and "C"  are not altered.
}
\description{
This function converts a data.frame of dissimilarity indices
(beta diversity) between sites to similarity indices.
}
\note{
The behaviour of this function changes depending on column names. Columns
"Site1" and "Site2" are copied identically. If there are columns called
"a", "b", "c", "A", "B", "C", they will also be copied identically.

If a column is called "Euclidean", its similarity will be calculated based
on the following formula:

\eqn{Euclidean similarity = (1 - Euclidean distance) / Euclidean distance}

Otherwise, all other columns will be transformed into similarity with the
following formula:

\eqn{similarity = 1 - dissimilarity}
}
\examples{
\dontrun{
comat <- matrix(sample(0:1000, size = 50, replace = TRUE, prob = 1 / 1:1001), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

simil <- similarity(comat, metric = "all")
simil

dissimilarity <- similarity_to_dissimilarity(simil)
dissimilarity

simil <- dissimilarity_to_similarity(dissimilarity)
simil
}

}
\seealso{
\code{\link[=dissimilarity_to_similarity]{dissimilarity_to_similarity()}}
}
\author{
Boris Leroy (\email{leroy.boris@gmail.com}),
Maxime Lenormand (\email{maxime.lenormand@inrae.fr}) and
Pierre Denelle (\email{pierre.denelle@gmail.com})
}
