% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fishsf}
\alias{fishsf}
\title{Spatial distribution of fish in Europe}
\format{
A
\describe{
\item{ID}{Unique site identifier}
\item{geometry}{Geometry of the site}
}
}
\usage{
fishsf
}
\description{
A dataset containing the geometry of the 338 sites.
}
\keyword{datasets}
