% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_binaries.R
\name{install_binaries}
\alias{install_binaries}
\title{Download, unzip, check permission and test the bioRgeo's executable binary files}
\usage{
install_binaries(binpath = NULL)
}
\arguments{
\item{binpath}{a \code{character} indicating the path to the folder that will 
host the 'bin' folder containing the executable binary files (bioRgeo's 
package by default, if you use a different folder please be sure to update 
the \code{binpath} in \link{netclu_infomap}, \link{netclu_louvain} and 
\link{netclu_oslom}).}
}
\description{
This function downloads and unzips the 'bin' folder needed to run some bioRgeo's 
functions. It also checks if the files have the permissions to be executed as 
programs. It finally tests if the executable binary files are running properly.
}
\author{
Maxime Lenormand (\email{maxime.lenormand@inrae.fr}),
Boris Leroy (\email{leroy.boris@gmail.com}) and
Pierre Denelle (\email{pierre.denelle@gmail.com}),
}
