% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_binaries.R
\name{install_binaries}
\alias{install_binaries}
\title{Download, unzip, check permission and test the bioRgeo's executables}
\usage{
install_binaries(binpath = NULL)
}
\arguments{
\item{binpath}{a string indicating the path to the folder that will host the bin folder (bioRgeo's package by default,
if you use a different folder please be sure to update the \code{binpath} in \link{infomap}, \link{louvain}
and \link{oslom})}
}
\description{
This function download and unzip the bin folder needed to run some bioRgeo's functions. It also checks if the files
have the permissions to be executed as programs. It finally tests if the executable are running properly.
}
\author{
Pierre Denelle (\email{pierre.denelle@gmail.com}),
Maxime Lenormand (\email{maxime.lenormand@inrae.fr}) and
Boris Leroy (\email{leroy.boris@gmail.com})
}
