% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_binaries.R
\name{install_binaries}
\alias{install_binaries}
\title{Download, unzip, check permission and test the bioregion's binary files}
\usage{
install_binaries(binpath = NULL, infomap_version = c("2.1.0", "2.6.0"))
}
\arguments{
\item{binpath}{a \code{character} indicating the path to the folder that will
host the 'bin' folder containing the binary files (bioregion's
package by default, if you use a different folder please be sure to update
the \code{binpath} in \link{netclu_infomap}, \link{netclu_louvain} and \link{netclu_oslom}).}

\item{infomap_version}{a \code{character} vector indicating the Infomap
version(s) to install.}
}
\description{
This function downloads and unzips the 'bin' folder needed to run some
functions of bioregion. It also checks if the files have the permissions to
be executed as programs. It finally tests if the binary files
are running properly.
}
\note{
Only the Infomap version 2.1.0 and 2.6.0 are available for now.
}
\author{
Maxime Lenormand (\email{maxime.lenormand@inrae.fr}),
Boris Leroy (\email{leroy.boris@gmail.com}) and
Pierre Denelle (\email{pierre.denelle@gmail.com})
}
