% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/louvain.R
\name{louvain}
\alias{louvain}
\title{Louvain community finding}
\usage{
louvain(
  net,
  weight = TRUE,
  lang = "Cpp",
  q = 0,
  c = 0.5,
  k = 1,
  delete_temp = TRUE,
  path_temp = "louvain_temp",
  binpath = NULL
)
}
\arguments{
\item{net}{a two- or three-column \code{data.frame} representing a network with the two first columns
as (un)directed links between pair of nodes and an optional third column indicating the weight of the link}

\item{weight}{a boolean indicating if the weights should be considered if there is a third column}

\item{lang}{a string indicating what version of Louvain should be used (igraph or Cpp, see Details)}

\item{q}{the quality function used to compute partition of the graph (modularity is chosen by default, see Details)}

\item{c}{the parameter for the Owsinski-Zadrozny quality function (between 0 and 1, 0.5 is chosen by default)}

\item{k}{the kappa_min value for the Shi-Malik quality function (it must be > 0, 1 is chosen by default)}

\item{delete_temp}{a boolean indicating if the temporary folder should be removed (see Details)}

\item{path_temp}{a string indicating the path to the temporary folder (see Details)}

\item{binpath}{a string indicating the path to the bin folder (see \link{bin} and Details)}
}
\value{
A \code{data.frame} providing one community by node.
}
\description{
This function finds communities in a (un)weighted undirected network based on the Louvain algorithm.
}
\details{
Louvain is a network community detection algorithm proposed in \insertCite{Blondel2008}{bioRgeo}. This function
proposed two implementations of the function (parameter \code{lang}): the \href{https://cran.r-project.org/web/packages/igraph/index.html}{igraph} implementation
(\link[igraph]{cluster_louvain}) and the C++ implementation (\url{https://sourceforge.net/projects/louvain/}, version 0.3).
The latest offers the possibility to choose among several quality functions, \code{q = 0} for the classical
Newman-Girvan criterion (also called "Modularity"), 1 for the Zahn-Condorcet criterion, 2 for
the Owsinski-Zadrozny criterion (you should specify the value of the parameter with option -c), 3	for the Goldberg
Density criterion, 4	for the A-weighted Condorcet criterion, 5 for the Deviation to Indetermination criterion, 6
for the Deviation to Uniformity criterion, 7 for the Profile Difference criterion, 8	for the Shi-Malik criterion
(you should specify the value of kappa_min with option -k) and 9	for the Balanced Modularity criterion.

The C++ version of Louvain is based on the version 0.3 (\url{https://sourceforge.net/projects/louvain/}).
This function needs executables files to run. They can be installed with \link{bin}. If you set the path to
the folder that will host the bin folder manually while running \link{bin} please make sure to set \code{binpath}
accordingly.

The C++ version of Louvain generates temporary folders and/or files that are stored in the \code{path_temp} folder
(folder "louvain_temp" in the workind directory by default). This temporary folder is removed by default
(\code{delete_temp = TRUE}).
}
\examples{
comat=matrix(sample(1000,50),5,10)
rownames(comat)=paste0("Site",1:5)
colnames(comat)=paste0("Species",1:10)

net=spproject(comat,metric="Simpson")
com=louvain(net, lang="igraph")
}
\references{
\insertRef{Blondel2008}{bioRgeo}
}
\seealso{
\link{infomap}, \link{oslom}
}
\author{
Pierre Denelle (\email{pierre.denelle@gmail.com}),
Maxime Lenormand (\email{maxime.lenormand@inrae.fr}) and
Boris Leroy (\email{leroy.boris@gmail.com})
}
