% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_to_df.R
\name{mat_to_net}
\alias{mat_to_net}
\title{Create a data.frame from a contingency table}
\usage{
mat_to_net(comat, weight = FALSE, remove_absent_objects = TRUE)
}
\arguments{
\item{comat}{a contingency table (i.e. \code{matrix})}

\item{weight}{a boolean indicating if the value are weights}

\item{remove_absent_objects}{a boolean determining whether absent
objects from the contingency table have to be removed from the output}
}
\value{
A \code{data.frame} where each row represents the interaction between
two objects and an optional third column indicating the weight of the interaction.
}
\description{
This function creates a two- or three-columns \code{data.frame} where
each row represents the interaction between two objects (site and species for example)
and an optional third column indicating the weight of the interaction (if \code{weight = TRUE})
from a contingency table (sites as rows and species as columns for example).
}
\examples{
comat=matrix(sample(1000,50),5,10)
rownames(comat)=paste0("Site",1:5)
colnames(comat)=paste0("Species",1:10)

df=mat_to_net(comat,weight=TRUE)
}
\seealso{
\link{net_to_mat}
}
\author{
Pierre Denelle (\email{pierre.denelle@gmail.com}),
Maxime Lenormand (\email{maxime.lenormand@inrae.fr}) and
Boris Leroy (\email{leroy.boris@gmail.com})
}
