% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netclu_infomap.R
\name{netclu_infomap}
\alias{netclu_infomap}
\title{Infomap community finding}
\usage{
netclu_infomap(
  net,
  weight = TRUE,
  index = names(net)[3],
  nbmod = 0,
  markovtime = 1,
  seed = 1,
  numtrials = 1,
  twolevel = FALSE,
  directed = FALSE,
  bipartite_version = FALSE,
  bipartite = FALSE,
  primary_col = 1,
  feature_col = 2,
  remove_feature = TRUE,
  delete_temp = TRUE,
  path_temp = "infomap_temp",
  binpath = NULL
)
}
\arguments{
\item{net}{a two- or three-column \code{data.frame} representing a network with the two first columns
as (un)directed links between pair of nodes and an optional third column indicating the weight of the link}

\item{weight}{a boolean indicating if the weights should be considered if there is a third column}

\item{index}{a \code{character} string providing the name of the dissimilarity
index to use, corresponding to the column
name in \code{dissimilarity}. By default, the third column name of
 \code{net} is used.}

\item{nbmod}{penalize solutions the more they differ from this number (0 by default for no preferred number of modules)}

\item{markovtime}{scales link flow to change the cost of moving between modules, higher values results
in fewer modules (default is 1)}

\item{seed}{for the random number generator}

\item{numtrials}{for the number of trials before picking up the best solution}

\item{twolevel}{a boolean indicating if the algorithm should optimize a two-level partition of the network
(default is multi-level)}

\item{directed}{a boolean indicating if the network is directed (from column 1 to column 2)}

\item{bipartite_version}{a boolean indicating if the bipartite version of Infomap should be used (see Details)}

\item{bipartite}{a boolean indicating if the network is bipartite (see Details)}

\item{primary_col}{name or number for the column of primary nodes (i.e. site)}

\item{feature_col}{name or number for the column of feature nodes (i.e. species)}

\item{remove_feature}{a boolean indicating if the feature nodes should be removed from the outputs (TRUE by default)}

\item{delete_temp}{a boolean indicating if the temporary folder should be removed (see Details)}

\item{path_temp}{a string indicating the path to the temporary folder (see Details)}

\item{binpath}{a string indicating the path to the bin folder (see \link{install_binaries} and Details)}
}
\value{
A \code{data.frame} providing one partition by hierarchical level.
}
\description{
This function finds communities in a (un)weighted (un)directed network based on the Infomap algorithm
(\url{https://github.com/mapequation/infomap}, version 1.6.0).
}
\details{
Infomap is a network clustering algorithm based on the Map equation proposed in
\insertCite{Rosvall2008}{bioRgeo} that finds communities in (un)weighted and (un)directed networks.
Infomap has two ways to deal with bipartite networks. The first possibility is to consider the bipartite network
as unipartite network (arguments \code{bipartite}, \code{primary_col}, \code{feature_col} and \code{remove_feature}).
The second possibility is to set the \code{bipartite_version} argument to TRUE in order to
approximate a two-step random walker (see \url{https://www.mapequation.org/infomap/} for more information).

This function is based on the 2.1.0 C++ version of Infomap (\url{https://github.com/mapequation/infomap/releases}).
This function needs executable files to run. They can be installed with \link{install_binaries}. If you set the path to
the folder that will host the bin folder manually while running \link{install_binaries} please make sure to set \code{binpath}
accordingly.

The C++ version of Infomap generates temporary folders and/or files that are stored in the \code{path_temp} folder
(folder "infomap_temp" in the working directory by default). This temporary folder is removed by default
(\code{delete_temp = TRUE}).
}
\examples{
comat <- matrix(sample(1000, 50), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

net <- similarity(comat, metric = "Simpson")
# com=netclu_infomap(net) # run install_binaries() to use this function
}
\references{
\insertRef{Rosvall2008}{bioRgeo}
}
\seealso{
\link{netclu_louvain}, \link{netclu_oslom}
}
\author{
Pierre Denelle (\email{pierre.denelle@gmail.com}),
Maxime Lenormand (\email{maxime.lenormand@inrae.fr}) and
Boris Leroy (\email{leroy.boris@gmail.com})
}
