% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netclu_infomap.R
\name{netclu_infomap}
\alias{netclu_infomap}
\title{Infomap community finding}
\usage{
netclu_infomap(
  net,
  weight = TRUE,
  index = names(net)[3],
  nbmod = 0,
  markovtime = 1,
  seed = 0,
  numtrials = 1,
  twolevel = FALSE,
  show_hierarchy = FALSE,
  directed = FALSE,
  bipartite_version = FALSE,
  bipartite = FALSE,
  site_col = 1,
  species_col = 2,
  return_node_type = "both",
  version = "2.7.1",
  binpath = "tempdir",
  path_temp = "infomap_temp",
  delete_temp = TRUE
)
}
\arguments{
\item{net}{the output object from \code{\link[=similarity]{similarity()}} or
\code{\link[=dissimilarity_to_similarity]{dissimilarity_to_similarity()}}.
If a \code{data.frame} is used, the first two columns represent pairs of
sites (or any pair of nodes), and the next column(s) are the similarity
indices.}

\item{weight}{a \code{boolean} indicating if the weights should be considered
if there are more than two columns.}

\item{index}{name or number of the column to use as weight. By default,
the third column name of \code{net} is used.}

\item{nbmod}{penalize solutions the more they differ from this number (0 by
default for no preferred number of modules).}

\item{markovtime}{scales link flow to change the cost of moving between
modules, higher values results in fewer modules (default is 1).}

\item{seed}{for the random number generator (0 for random by default).}

\item{numtrials}{for the number of trials before picking up the best
solution.}

\item{twolevel}{a \code{boolean} indicating if the algorithm should optimize a
two-level partition of the network (default is multi-level).}

\item{show_hierarchy}{a \code{boolean} specifying if the hierarchy of community
should be identifiable in the outputs (FALSE by default).}

\item{directed}{a \code{boolean} indicating if the network is directed (from
column 1 to column 2).}

\item{bipartite_version}{a \code{boolean} indicating if the bipartite version of
Infomap should be used (see Note).}

\item{bipartite}{a \code{boolean} indicating if the network is bipartite
(see Note).}

\item{site_col}{name or number for the column of site nodes
(i.e. primary nodes).}

\item{species_col}{name or number for the column of species nodes
(i.e. feature nodes).}

\item{return_node_type}{a \code{character} indicating what types of nodes
("sites", "species" or "both") should be returned in the output
(\code{return_node_type = "both"} by default).}

\item{version}{a \code{character} indicating the Infomap version to use.}

\item{binpath}{a \code{character} indicating the path to the bin folder
(see \link{install_binaries} and Details).}

\item{path_temp}{a \code{character} indicating the path to the temporary folder
(see Details).}

\item{delete_temp}{a \code{boolean} indicating if the temporary folder should
be removed (see Details).}
}
\value{
A \code{list} of class \code{bioregion.clusters} with five slots:
\enumerate{
\item{\strong{name}: \code{character} containing the name of the algorithm}
\item{\strong{args}: \code{list} of input arguments as provided by the user}
\item{\strong{inputs}: \code{list} of characteristics of the clustering process}
\item{\strong{algorithm}: \code{list} of all objects associated with the
clustering procedure, such as original cluster objects}
\item{\strong{clusters}: \code{data.frame} containing the clustering results}}

In the \code{algorithm} slot, users can find the following elements:

\itemize{
\item{\code{cmd}: the command line use to run Infomap}
\item{\code{version}: the Infomap version}
\item{\code{web}: Infomap's GitHub repository}
}
}
\description{
This function finds communities in a (un)weighted (un)directed network based
on the Infomap algorithm (\url{https://github.com/mapequation/infomap}).
}
\details{
Infomap is a network clustering algorithm based on the Map equation proposed
in \insertCite{Rosvall2008}{bioregion} that finds communities in (un)weighted
and (un)directed networks.

This function is based on the C++ version of Infomap
(\url{https://github.com/mapequation/infomap/releases}).
This function needs binary files to run. They can be installed
with \link{install_binaries}.

\strong{If you changed the default path to the \code{bin} folder
while running \link{install_binaries} PLEASE MAKE SURE to set \code{binpath}
accordingly.}

The C++ version of Infomap generates temporary folders and/or files that are
stored in the \code{path_temp} folder ("infomap_temp" with an unique timestamp
located in the bin folder in \code{binpath} by default). This temporary folder is
removed by default (\code{delete_temp = TRUE}).

Several version of Infomap are available in the package. See
\link{install_binaries} for more details.
}
\note{
Infomap has been designed to deal with bipartite networks. To use this
functionality set the \code{bipartite_version} argument to TRUE in order to
approximate a two-step random walker (see
\url{https://www.mapequation.org/infomap/} for more information). Note that
a bipartite network can also be considered as unipartite network
(\code{bipartite = TRUE}).

In both cases do not forget to indicate which of the first two columns is
dedicated to the site nodes (i.e. primary nodes) and species nodes (i.e.
feature nodes) using the arguments \code{site_col} and \code{species_col}.
The type of nodes returned in the output can be chosen with the argument
\code{return_node_type} equal to \code{"both"} to keep both types of nodes, \code{"sites"}
to preserve only the sites nodes and \code{"species"} to preserve only the
species nodes.
}
\examples{
comat <- matrix(sample(1000, 50), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

net <- similarity(comat, metric = "Simpson")
com <- netclu_infomap(net)

}
\references{
\insertRef{Rosvall2008}{bioregion}
}
\seealso{
\link{install_binaries}, \link{netclu_louvain}, \link{netclu_oslom}
}
\author{
Maxime Lenormand (\email{maxime.lenormand@inrae.fr}),
Pierre Denelle (\email{pierre.denelle@gmail.com}) and
Boris Leroy (\email{leroy.boris@gmail.com})
}
