% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhclu_kmeans.R
\name{nhclu_kmeans}
\alias{nhclu_kmeans}
\title{Non hierarchical clustering: k-means analysis}
\usage{
nhclu_kmeans(
  dissimilarity,
  index = names(dissimilarity)[3],
  n_clust = NULL,
  iter_max = 10,
  nstart = 10,
  algorithm = "Hartigan-Wong"
)
}
\arguments{
\item{dissimilarity}{the output object from \code{\link{similarity_to_dissimilarity}},
a \code{data.frame} with the first columns called "Site1" and "Site2", and
the other columns being the dissimilarity indices or a \code{dist} object}

\item{index}{a \code{character} string providing the name of the dissimilarity
index to use, corresponding to the column
name in \code{dissimilarity}. By default, the third column name of
 \code{dissimilarity} is used.}

\item{n_clust}{an \code{integer} or a \code{vector} of \code{integers}
specifying the requested number(s) of clusters}

\item{iter_max}{an \code{integer} specifying the maximum number of
iterations for the kmeans method (see \link[stats:kmeans]{stats::kmeans()})}

\item{nstart}{an \code{integer} specifying how many random sets of
\code{n_clust} should be selected as starting points for the kmeans analysis
(see \link[stats:kmeans]{stats::kmeans()})}

\item{algorithm}{a \code{character string} specifying the algorithm to use for
kmean (see \link[stats:kmeans]{stats::kmeans()})}
}
\value{
A \code{list} of class \code{bioRgeo.clusters} with five slots:
\enumerate{
\item{\bold{name}: \code{character string} containing the name of the algorihtm}
\item{\bold{args}: \code{list} of input arguments as provided by the user}
\item{\bold{inputs}: \code{list} of characteristics of the input dataset}
\item{\bold{algorithm}: \code{list} of all objects associated with the
 clustering procedure, such as original cluster objects}
\item{\bold{clusters}: \code{data.frame} containing the clustering results}}
}
\description{
This function performs non hierarchical
clustering on the basis of dissimilarity with a k-means analysis.
}
\details{
This method partitions the data into k groups
 such that that the sum of squares of euclidean distances from points to the
 assigned cluster centres is minimized. k-means cannot be applied directly
 on dissimilarity/beta-diversity metrics, because these distances are not
 euclidean. Therefore, it requires first to transform the dissimilarity matrix
 with a Principal Coordinate Analysis (using the function
 \link[ape:pcoa]{ape::pcoa()}), and then applying k-means on the coordinates
 of points in the PCoA. Because this makes an additional transformation of
 the initial matrix of dissimilarity, the partitioning around medoids method
 should be prefered (\code{\link{nhclu_pam}})
}
\examples{
simil <- similarity(vegemat, metric = "all")
dissimilarity <- similarity_to_dissimilarity(simil)

clust1 <- nhclu_kmeans(dissimilarity,
    n_clust = 2:10,
    index = "Simpson")
clust2 <- nhclu_kmeans(dissimilarity,
    n_clust = 2:25,
    index = "Simpson")
partition_metrics(clust2,
                  dissimilarity = dissimilarity,
                  eval_metric = "pc_distance",
                  partition_optimisation = TRUE)
partition_metrics(clust2,
                  sp_site_table = vegemat,
                  eval_metric = "avg_endemism",
                  partition_optimisation = TRUE)
}
