% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhclu_pam.R
\name{nhclu_pam}
\alias{nhclu_pam}
\title{Non hierarchical clustering: partitioning around medoids}
\usage{
nhclu_pam(
  dissimilarity,
  index = names(dissimilarity)[3],
  n_clust = NULL,
  nstart = if (variant == "faster") 1 else NA,
  variant = "faster",
  cluster_only = FALSE,
  ...
)
}
\arguments{
\item{dissimilarity}{the output object from \code{\link{dissimilarity}} or
 \code{\link{similarity_to_dissimilarity}}, or a \code{dist} object. 
 If a \code{data.frame} is used, the first two 
columns represent pairs of sites (or any pair of nodes), and the next column(s)
are the dissimilarity indices.}

\item{index}{name or number of the dissimilarity column to use. By default, 
the third column name of
 \code{dissimilarity} is used.}

\item{n_clust}{an \code{integer} or a \code{vector} of \code{integers}
specifying the requested number(s) of clusters}

\item{nstart}{an \code{integer} specifying the number of random “starts”
for the pam algorithm. By default, 1 (for the \code{"faster"} variant)}

\item{variant}{a \code{character} string specifying the variant of pam
to use, by default "faster". See \link[cluster:pam]{cluster::pam()} for
more details}

\item{cluster_only}{a \code{boolean} specifying if only the clustering
should be returned from the \link[cluster:pam]{cluster::pam()} function
(more efficient)}

\item{...}{you can add here further arguments to be passed to \code{pam()}
(see \link[cluster:pam]{cluster::pam()})}
}
\value{
A \code{list} of class \code{bioRgeo.clusters} with five slots:
\enumerate{
\item{\bold{name}: \code{character string} containing the name of the algorihtm}
\item{\bold{args}: \code{list} of input arguments as provided by the user}
\item{\bold{inputs}: \code{list} of characteristics of the input dataset}
\item{\bold{algorithm}: \code{list} of all objects associated with the
 clustering procedure, such as original cluster objects}
\item{\bold{clusters}: \code{data.frame} containing the clustering results}}
}
\description{
This function performs non hierarchical
clustering on the basis of dissimilarity with partitioning around medoids.
}
\details{
This method partitions data into
 the chosen number of cluster on the basis of the input dissimilarity matrix.
 It is more robust than k-means because it minimizes the sum of dissimilarity
 between cluster centres and points assigned to the cluster -
 whereas the k-means approach minimizes the sum of squared euclidean
 distances (thus k-means cannot be applied directly on the input dissimilarity
 matrix if the distances are not euclidean).
}
\examples{
dissim <- dissimilarity(vegemat, metric = "all")

clust1 <- nhclu_pam(dissim,
     n_clust = 2:10,
     index = "Simpson")
clust2 <- nhclu_pam(dissim,
     n_clust = 2:25,
     index = "Simpson")
partition_metrics(clust2,
                   dissimilarity = dissim,
                   eval_metric = "pc_distance",
                   partition_optimisation = TRUE)
\donttest{                  
partition_metrics(clust2,
                   sp_site_table = vegemat,
                  eval_metric = "avg_endemism",
                   partition_optimisation = TRUE)
                   }
}
\seealso{
\link{nhclu_kmeans}
}
\author{
Boris Leroy (\email{leroy.boris@gmail.com}),
Pierre Denelle (\email{pierre.denelle@gmail.com}) and
Maxime Lenormand (\email{maxime.lenormand@inrae.fr})
}
