% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_species_metrics.R
\name{site_species_metrics}
\alias{site_species_metrics}
\title{Calculate contribution metrics of sites and species to each bioregion}
\usage{
site_species_metrics(
  bioregionalization,
  bioregion_indices = c("Specificity", "NSpecificity", "Fidelity", "IndVal", "NIndVal",
    "Rho", "CoreTerms"),
  bioregionalization_indices = "P",
  data_type = "auto",
  node_type = "site",
  comat,
  similarity = NULL,
  index = names(similarity)[3],
  verbose = TRUE
)
}
\arguments{
\item{bioregionalization}{A \code{bioregion.clusters} object.}

\item{bioregion_indices}{A \code{character} vector or a single \code{character} string
specifying the indices to compute. Several indices belonging to different
categories are available: species to bioregions (\code{"Specificity"},
\code{"NSpecificity"}, \code{"Fidelity"}, \code{"NFidelity"}, \code{"IndVal"}, \code{"NIndVal"},
\code{"Rho"} and \code{"CoreTerms"}), site to bioregions (\code{"MeanSim"} and
\code{"SdSim"}), or site to species clusters when a cluster/bioregion has been
assigned to the species  in \code{bioregionalization} (see Details). If \code{"all"}
is specified (default), all indices will be calculated.}

\item{bioregionalization_indices}{A \code{character} vector or a single \code{character}
string specifying the bioregionalization indices to compute. Some aggregated
indices (such as the participation coefficient or Silhouette index) can be
derived from the \code{bioregion_indices} (see Details). If \code{"all"} is specified
(default), all bioregionalization indices will be calculated.}

\item{data_type}{A \code{character} string specifying which type of data should
be considered to compute the related indices (\code{"A"}, \code{"B"},
\code{"IndVal"} and \code{"Rho"}): occurrences or abundances. By default (\code{"auto"}),
the type of data is inferred from \code{bioregionalization} and/or the provided
co-occurrence matrix (argument \code{comat}). Other available options are
\code{"occurence"}, \code{"abundance"} or \code{"both"} (see Details).}

\item{node_type}{A \code{character} string specifying whether the related
indices (\code{"Specificity"}, \code{"NSpecificity"}, \code{"Fidelity"}, \code{"NFidelity"},
\code{"IndVal"}, \code{"NIndVal"}, \code{"Rho"} and \code{"CoreTerms"})
should be computed as contributions from species to bioregions (\code{"site"} by
default), from sites to species clusters (\code{"species"}), or for \code{"both"}.
The latter type of contribution is only available if a cluster has been
assigned to the species in \code{bioregionalization} (see Details).}

\item{comat}{A co-occurrence \code{matrix} with sites as rows and species as
columns.}

\item{similarity}{The output object from \code{\link[=similarity]{similarity()}} or
\code{\link[=dissimilarity_to_similarity]{dissimilarity_to_similarity()}}.}

\item{index}{The name or number of the column to use as similarity.
By default, the third column name of \code{similarity} is used.}

\item{verbose}{A \code{boolean} indicating whether to
display progress messages. Set to \code{FALSE} to suppress these messages.}
}
\value{
A \code{list} containing between one and six elements (listed below), depending on
the selected indices (\code{bioregion_indices} and/or \code{bioregionalization_indices})
and the type of nodes (\code{site} and/or \code{species}).
\itemize{
\item{\strong{species_bioregions}: A \code{data.frame} containing the
species-to-bioregions indice(s) based on \code{comat}.}
\item{\strong{species_bioregionalization}: A \code{data.frame} containing the
species-to-bioregionalization indice(s) based on \code{comat}.}
\item{\strong{site_clusters}: A \code{data.frame} containing the
site-to-species clusters indice(s) based on \code{comat}.}
\item{\strong{site_clustering}: A \code{data.frame} containing the
site-to-species clustering indice(s) based on \code{comat}.}
\item{\strong{site_bioregions}: A \code{data.frame} containing the
site-to-bioregions indice(s) based on \code{similarity}.}
\item{\strong{site_bioregionalization}: A \code{data.frame} containing the
site-to-bioregionalization indice(s) based on \code{similarity}.}
}

Note that if \code{bioregionalization} contains more than one partition
(i.e., if \code{dim(bioregionalization$clusters) > 2}), a list of lists will be
returned, with one sublist per partition.
}
\description{
This function computes metrics to assess the contribution of a given
species or site to each bioregion.
}
\details{
The \strong{first type} of indices provided by this function is based on the
contribution of each species to a given bioregion (\code{node_type = "site"} or
\code{node_type = "both"}). This is calculated from the bioregion assigned to
each site (as defined in \code{bioregionalization}) and
an associated site–species co-occurrence matrix \code{comat} (a strict match
between site IDs is required).

Occurrence-based indices (\code{data_type = "occurrence"} or \code{data_type = "both"})
are derived from three core terms:
\itemize{
\item \strong{n_sb}: Number of sites belonging to bioregion \strong{b} in which species
\strong{s} is present.
\item \strong{n_s}: Total number of sites in which species \strong{s} is present.
\item \strong{n_b}: Total number of sites belonging to bioregion \strong{b}.
}

These \strong{species-to-bioregion} indices include:
\itemize{
\item \href{https://biorgeo.github.io/bioregion/articles/a5_2_summary_metrics.html#specificity-occurrence}{\strong{Specificity}},
as described in De Cáceres M & Legendre P (2009).
\item \href{https://biorgeo.github.io/bioregion/articles/a5_2_summary_metrics.html#npecificity-occurrence}{\strong{NSpecificity}}
is the normalized version of \strong{Specificity} accounting for bioregion size,
as described in De Cáceres M & Legendre P (2009).
\item \href{https://biorgeo.github.io/bioregion/articles/a5_2_summary_metrics.html#fidelity-occurrence}{\strong{Fidelity}}
as described in De Cáceres M & Legendre P (2009).
\item \href{https://biorgeo.github.io/bioregion/articles/a5_2_summary_metrics.html#indval-occurrence}{\strong{IndVal}},
as described in De Cáceres M & Legendre P (2009).
\item \href{https://biorgeo.github.io/bioregion/articles/a5_2_summary_metrics.html#nindval-occurrence}{\strong{NIndVal}}
is the normalized version of \strong{IndVal} accounting for bioregion size,
as described in De Cáceres M & Legendre P (2009).
\item \href{https://biorgeo.github.io/bioregion/articles/a5_2_summary_metrics.html#rho-occurrence}{\strong{Rho}},
as described in Lenormand \emph{et al.} (2019).
}

\strong{Species-to-bioregionalization} indices can also be computed, such as:
\itemize{
\item \href{https://biorgeo.github.io/bioregion/articles/a5_2_summary_metrics.html#participation-occurrence}{\strong{Participation}},
as described in Denelle \emph{et al.} (2020).
}

Abundance-weighted versions of these indices (\code{data_type = "abundance"} or
\code{data_type = "both"}) can also be derived using the following analogous core
terms:
\itemize{
\item \strong{w_sb}: Sum of abundances of species \strong{s} in sites of bioregion \strong{b}.
\item \strong{w_s}: Total abundance of species \strong{s}.
\item \strong{w_b}: Total abundance of all species present in sites of bioregion \strong{b}.
}

These abundance-weighted terms correspond directly to their occurrence-based
counterparts and allow computing abundance versions of the same
indices. Detailed formulas and examples are provided in the package vignette.

The \strong{second type} of indices provided by this function is based on the
contribution of each site to a given \emph{species cluster} (\code{node_type = "species"}
or \code{node_type = "both"}, only when a
cluster or bioregion has been assigned to species in \code{bioregionalization}).
This is calculated from the cluster assigned to each species (as defined in
\code{bioregionalization}) and an associated site–species co-occurrence matrix
(a strict match between species IDs is required).

In this case, occurrence-based indices (\code{data_type = "occurrence"} or
\code{data_type = "both"}) are derived from three core terms:
\itemize{
\item \strong{n_gc}: Number of species belonging to cluster \strong{c} that are present
in site \strong{g}.
\item \strong{n_g}: Total number of species present in site \strong{g}.
\item \strong{n_c}: Total number of species belonging to cluster \strong{c}.
}

As for the first type, all indices (including their abundance-weighted
versions when \code{data_type = "abundance"} or \code{data_type = "both"}) can be
derived from these \strong{site-to-clusters} relationships. Further
details, mathematical definitions, and examples are provided in the package
vignette.

The \strong{third type} of indices provided by this function, not included by
default, is based on the contribution of each site to a given bioregion. This
is calculated from the bioregion assigned to each site (as defined in
\code{bioregionalization}) and a site–site similarity metric (\code{similarity}) (a
strict match between site IDs is required).

These \strong{site-to-bioregion} indices include:
\itemize{
\item \href{https://biorgeo.github.io/bioregion/articles/a5_2_summary_metrics.html#meansim}{\strong{MeanSim}}:
The mean similarity of each site to the sites of every bioregion.
\item \href{https://biorgeo.github.io/bioregion/articles/a5_2_summary_metrics.html#sdsim}{\strong{SdSim}}:
The corresponding standard deviation of similarity values.
}

\strong{Site-to-bioregionalization} indices can also be computed, such as:
\itemize{
\item \href{https://biorgeo.github.io/bioregion/articles/a5_2_summary_metrics.html#silhouette}{\strong{Silhouette}},
as described in Rousseeuw (1987) (similarity-based version).
}
}
\note{
If \code{data_type = "auto"}, the choice between occurrence- or abundance-
based indices will be determined automatically from the input data, and a
message will explain the choice made.
}
\examples{
data(fishmat)

fishsim <- similarity(fishmat, metric = "Jaccard")

bioregionalization <- hclu_hierarclust(similarity_to_dissimilarity(fishsim),
                                       index = "Jaccard",
                                       method = "average",
                                       randomize = TRUE,
                                       optimal_tree_method = "best",
                                       n_clust = c(1,2,3),
                                       verbose = FALSE)
                                     
ind <- site_species_metrics(bioregionalization = bioregionalization,
                             bioregion_indices = "all",
                             bioregionalization_indices = "all",
                             data_type = "auto",
                             node_type = "site",
                             comat = fishmat,
                             similarity = fishsim,
                             index = 3,
                             verbose = TRUE)
                                     
}
\references{
De Cáceres M & Legendre P (2009) Associations between species and groups of
sites: indices and statistical inference. \emph{Ecology} 90, 3566--3574.

Denelle P, Violle C & Munoz F (2020) Generalist plants are more competitive
and more functionally similar to each other than specialist plants: insights
from network analyses. \emph{Journal of Biogeography} 47, 1922–-1933.

Lenormand M, Papuga G, Argagnon O, Soubeyrand M, Alleaume S & Luque S (2019)
Biogeographical network analysis of plant species distribution in the
Mediterranean region. \emph{Ecology and Evolution} 9, 237--250.

Rousseeuw PJ (1987) Silhouettes: A graphical aid to the interpretation and
validation of cluster analysis. \emph{Journal of Computational and Applied
Mathematics} 20, 53--65.
}
\seealso{
For more details illustrated with a practical example,
see the vignette:
\url{https://biorgeo.github.io/bioregion/articles/a5_2_summary_metrics.html}.

Associated functions:
\link{bioregion_metrics} \link{bioregionalization_metrics}
}
\author{
Maxime Lenormand (\email{maxime.lenormand@inrae.fr}) \cr
Boris Leroy (\email{leroy.boris@gmail.com}) \cr
Pierre Denelle (\email{pierre.denelle@gmail.com})
}
