% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_species_metrics.R
\name{site_species_metrics}
\alias{site_species_metrics}
\title{Calculate contribution metrics of sites and species}
\usage{
site_species_metrics(
  cluster_object,
  comat,
  indices = c("rho"),
  bipartite_link = NULL
)
}
\arguments{
\item{cluster_object}{a \code{bioregion.clusters} object or a \code{data.frame} or a
list of \code{data.frame} containing multiple partitions. At least two partitions
are required. If a list of \code{data.frame} is provided, they should all have
the same number of rows (i.e., same items in the clustering for all
partitions).}

\item{comat}{a co-occurrence \code{matrix} with sites as rows and species as
columns.}

\item{indices}{a \code{character} specifying the contribution metric to compute.
Available options are \code{rho} and \code{Cz}.}

\item{bipartite_link}{\code{NULL} by default. Needed for \code{Cz} indices. A
\code{data.frame} where each row represents the interaction between two nodes
and an optional third column indicating the weight of the interaction.}
}
\value{
A \code{data.frame} with the columns Bioregion, Species, and the desired summary
statistics, or a list of \code{data.frames} if \code{Cz} and other indices are
selected.
}
\description{
This function calculates metrics that assess the contribution of a given
species or site to its bioregion.
}
\details{
The \eqn{\rho} metric is derived from \insertRef{Lenormand2019}{bioregion}.
Its formula is the following:
\eqn{\rho_{ij} = (n_ij - ((n_i n_j)/n))/(sqrt(((n - n_j)/(n-1)) (1-(n_j/n)) ((n_i n_j)/n)))}

with \eqn{n} the number of sites, \eqn{n_i} the number of sites in which
species \eqn{i} is present, \eqn{n_j} the number of sites belonging to the
bioregion \eqn{j}, \eqn{n_ij} the number of occurrences of species \eqn{i}
in sites belonging to the bioregion \eqn{j}.

Affinity, fidelity and individual contributions describe how species are
linked to their bioregions.
Affinity of species to their region, \eqn{A_i = R_i / Z}
where \eqn{R_i} is the occurrence/range size of species \eqn{i} in its
associated cluster, and \eqn{Z} the total size (number of sites) of the
bioregion.

A high affinity means that the species is occupying most sites of its
associated bioregion.

Fidelity of species to their region, \eqn{F_i = R_i / D_i}
where \eqn{R_i} is the occurrence/range size of species \eqn{i} in its
associated bioregion, and \eqn{D_i} is its total occurrence/range size.

A high fidelity means that the species is not present in other bioregions
than their associated one.

Indicator Value of species, \eqn{IndVal = F_i * A_i}

\code{Cz} metrics are derived from \insertRef{Guimera2005}{bioregion}.
Their respective formula are:
\eqn{C_i = 1 - \sum_{s=1}^{N_M}{{(\frac{k_is}{k_i}})^2}}

where \eqn{k_{is}} is the number of links of node (species or site) \eqn{i}
to nodes in bioregion \eqn{s}, and \eqn{k_i} is the total degree of node
\eqn{i}. The participation coefficient of a node is therefore close to 1 if
its links are uniformly distributed among all the bioregions and 0 if all
its links are within its own bioregion.

And:
\eqn{z_i = \frac{k_i - \overline{k_{si}}}{\sigma_{k_{si}}}}

where \eqn{k_i} is the number of links of node (species or site) \eqn{i} to
other nodes in its bioregion \eqn{s_i}, \eqn{\overline{k_{si}}} is the
average of \eqn{k} over all the nodes in \eqn{s_i}, and
\eqn{\sigma_{k_{si}}} is the standard deviation of \eqn{k} in \eqn{s_i}.
The within-bioregion degree z-score measures how well-connected node \eqn{i}
is to other nodes in the bioregion.
}
\examples{
comat <- matrix(sample(1000, 50), 5, 10)
rownames(comat) <- paste0("Site", 1:5)
colnames(comat) <- paste0("Species", 1:10)

comat <- matrix(sample(0:1000, size = 500, replace = TRUE, prob = 1/1:1001),
                20, 25)
rownames(comat) <- paste0("Site",1:20)
colnames(comat) <- paste0("Species",1:25)

dissim <- dissimilarity(comat, metric = "Simpson")
clust1 <- nhclu_kmeans(dissim, n_clust = 3, index = "Simpson")

net <- similarity(comat, metric = "Simpson")
com <- netclu_greedy(net)

site_species_metrics(cluster_object = clust1, comat = comat,
indices = "rho")

# Contribution metrics
site_species_metrics(cluster_object = com, comat = comat,
indices = c("rho", "affinity", "fidelity", "indicator_value"))

# Cz indices
net_bip <- mat_to_net(comat, weight = TRUE)
clust_bip <- netclu_greedy(net_bip, bipartite = TRUE)
site_species_metrics(cluster_object = clust_bip, comat = comat, 
bipartite_link = net_bip, indices = "Cz")

}
\seealso{
\link{bioregionalization_metrics}
}
\author{
Pierre Denelle (\email{pierre.denelle@gmail.com}),
Boris Leroy (\email{leroy.boris@gmail.com}) and
Maxime Lenormand (\email{maxime.lenormand@inrae.fr})
}
