% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vegedf}
\alias{vegedf}
\title{Spatial distribution of Mediterranean vegetation (data.frame)}
\format{
A \code{data.frame} with 460,878 rows and 3 columns:
\describe{
  \item{Site}{Unique site identifier (corresponding to the field ID of vegesp).}
  \item{Species}{Unique species identifier.}
  \item{Abundance}{Species abundance}
}
}
\source{
\url{https://onlinelibrary.wiley.com/doi/full/10.1002/ece3.4718}
}
\usage{
vegedf
}
\description{
A dataset containing the abundance of 3,697 species in 715 sites.
}
\keyword{datasets}
