% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vegemat}
\alias{vegemat}
\title{Spatial distribution of Mediterrean Vegetation (co-occurence matrix)}
\format{
A co-occurence \code{matrix} with sites as rows and species as columns. Each element of the matrix
represents the abundance of the species in the site.
}
\source{
\href{https://onlinelibrary.wiley.com/doi/full/10.1002/ece3.4718}(https://onlinelibrary.wiley.com/doi/full/10.1002/ece3.4718)
}
\usage{
vegemat
}
\description{
A dataset containing the abundance of each of the 3,697 species in each of the 715 sites.
}
\keyword{datasets}
