% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vegesf}
\alias{vegesf}
\title{Spatial distribution of Mediterranean vegetation (spatial grid)}
\format{
A
\describe{
\item{ID}{Unique site identifier}
\item{geometry}{Geometry of the site}
}
}
\source{
\doi{10.1002/ece3.4718}
}
\usage{
vegesf
}
\description{
A dataset containing the geometry of the 715 sites.
}
\keyword{datasets}
