% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vegesp}
\alias{vegesp}
\title{Spatial distribution of Mediterrean Vegetation (spatial grid)}
\format{
A
\describe{
  \item{ID}{Unique site identifier.}
  \item{geometry}{Geometry of the site.}
}
}
\source{
\url{https://onlinelibrary.wiley.com/doi/full/10.1002/ece3.4718}
}
\usage{
vegesp
}
\description{
A dataset containing the geometry of the 715 sites.
}
\keyword{datasets}
